Compute melting temperature of multiple nucleic acid duplexes in batch

Compute the enthalpy and entropy of helix-coil transition, and then the melting temperature of multiple nucleic acid duplexes in batch.

meltingBatch(
  sequence,
  comp.sequence = NULL,
  environment.out = TRUE,
  options.out = TRUE,
  message.out = TRUE,
  ...
)

Arguments

sequence

A character vector of 5' to 3' sequences of one strand of the nucleic acid duplex (Note: Uridine and thymidine are not considered as identical).

comp.sequence

A character vector of 3' to 5' complementary sequences of the nucleic acid duplex. Complementary sequences are computed by default, but need to be specified in case of mismatches, inosine(s) or hydroxyadenine(s) between the two strands.

environment.out

logical. If TRUE, gives the melting temperature computation environment details in the output. Default is TRUE.

options.out

logical. If TRUE, gives the details about the options (default or user specified) used for melting temperature computation in the output. Default is TRUE.

message.out

logical. If TRUE, gives the error and/or warning messages, if any in the output. Default is TRUE.

...

Arguments for melting temperature computation (See melting).

Value

A data frame of the melting temperature computation results along with the details of environment, options and messages if specified by the arguments environment.out, options.out and message.out

respectively.

See also

Examples

sequence <- c("CAAAAAG", "CAAAAAAG", "TTTTATAATAAA", "CCATCGCTACC",
              "CAAACAAAG", "CCATTGCTACC", "CAAAAAAAG", "GTGAAC", "AAAAAAAA",
              "CAACTTGATATTATTA", "CAAATAAAG", "GCGAGC", "GGGACC",
              "CAAAGAAAG", "CTGACAAGTGTC", "GCGAAAAGCG")

meltingBatch(sequence, nucleic.acid.conc = 0.0004,
             hybridisation.type = "dnadna", Na.conc = 1)
#>       Environment.Sequence Environment.Complementary sequence
#>  [1,] "CAAAAAG"            "GTTTTTC"                         
#>  [2,] "CAAAAAAG"           "GTTTTTTC"                        
#>  [3,] "TTTTATAATAAA"       "AAAATATTATTT"                    
#>  [4,] "CCATCGCTACC"        "GGTAGCGATGG"                     
#>  [5,] "CAAACAAAG"          "GTTTGTTTC"                       
#>  [6,] "CCATTGCTACC"        "GGTAACGATGG"                     
#>  [7,] "CAAAAAAAG"          "GTTTTTTTC"                       
#>  [8,] "GTGAAC"             "CACTTG"                          
#>  [9,] "AAAAAAAA"           "TTTTTTTT"                        
#> [10,] "CAACTTGATATTATTA"   "GTTGAACTATAATAAT"                
#> [11,] "CAAATAAAG"          "GTTTATTTC"                       
#> [12,] "GCGAGC"             "CGCTCG"                          
#> [13,] "GGGACC"             "CCCTGG"                          
#> [14,] "CAAAGAAAG"          "GTTTCTTTC"                       
#> [15,] "CTGACAAGTGTC"       "GACTGTTCACAG"                    
#> [16,] "GCGAAAAGCG"         "CGCTTTTCGC"                      
#>       Environment.Nucleic acid concentration (M) Environment.Hybridization type
#>  [1,] "4e-04"                                    "dnadna"                      
#>  [2,] "4e-04"                                    "dnadna"                      
#>  [3,] "4e-04"                                    "dnadna"                      
#>  [4,] "4e-04"                                    "dnadna"                      
#>  [5,] "4e-04"                                    "dnadna"                      
#>  [6,] "4e-04"                                    "dnadna"                      
#>  [7,] "4e-04"                                    "dnadna"                      
#>  [8,] "4e-04"                                    "dnadna"                      
#>  [9,] "4e-04"                                    "dnadna"                      
#> [10,] "4e-04"                                    "dnadna"                      
#> [11,] "4e-04"                                    "dnadna"                      
#> [12,] "4e-04"                                    "dnadna"                      
#> [13,] "4e-04"                                    "dnadna"                      
#> [14,] "4e-04"                                    "dnadna"                      
#> [15,] "4e-04"                                    "dnadna"                      
#> [16,] "4e-04"                                    "dnadna"                      
#>       Environment.Na concentration (M) Environment.Mg concentration (M)
#>  [1,] "1"                              "0"                             
#>  [2,] "1"                              "0"                             
#>  [3,] "1"                              "0"                             
#>  [4,] "1"                              "0"                             
#>  [5,] "1"                              "0"                             
#>  [6,] "1"                              "0"                             
#>  [7,] "1"                              "0"                             
#>  [8,] "1"                              "0"                             
#>  [9,] "1"                              "0"                             
#> [10,] "1"                              "0"                             
#> [11,] "1"                              "0"                             
#> [12,] "1"                              "0"                             
#> [13,] "1"                              "0"                             
#> [14,] "1"                              "0"                             
#> [15,] "1"                              "0"                             
#> [16,] "1"                              "0"                             
#>       Environment.Tris concentration (M) Environment.K concentration (M)
#>  [1,] "0"                                "0"                            
#>  [2,] "0"                                "0"                            
#>  [3,] "0"                                "0"                            
#>  [4,] "0"                                "0"                            
#>  [5,] "0"                                "0"                            
#>  [6,] "0"                                "0"                            
#>  [7,] "0"                                "0"                            
#>  [8,] "0"                                "0"                            
#>  [9,] "0"                                "0"                            
#> [10,] "0"                                "0"                            
#> [11,] "0"                                "0"                            
#> [12,] "0"                                "0"                            
#> [13,] "0"                                "0"                            
#> [14,] "0"                                "0"                            
#> [15,] "0"                                "0"                            
#> [16,] "0"                                "0"                            
#>       Environment.dNTP concentration (M) Environment.DMSO concentration (%)
#>  [1,] "0"                                "0"                               
#>  [2,] "0"                                "0"                               
#>  [3,] "0"                                "0"                               
#>  [4,] "0"                                "0"                               
#>  [5,] "0"                                "0"                               
#>  [6,] "0"                                "0"                               
#>  [7,] "0"                                "0"                               
#>  [8,] "0"                                "0"                               
#>  [9,] "0"                                "0"                               
#> [10,] "0"                                "0"                               
#> [11,] "0"                                "0"                               
#> [12,] "0"                                "0"                               
#> [13,] "0"                                "0"                               
#> [14,] "0"                                "0"                               
#> [15,] "0"                                "0"                               
#> [16,] "0"                                "0"                               
#>       Environment.Formamide concentration (M or %)
#>  [1,] "0"                                         
#>  [2,] "0"                                         
#>  [3,] "0"                                         
#>  [4,] "0"                                         
#>  [5,] "0"                                         
#>  [6,] "0"                                         
#>  [7,] "0"                                         
#>  [8,] "0"                                         
#>  [9,] "0"                                         
#> [10,] "0"                                         
#> [11,] "0"                                         
#> [12,] "0"                                         
#> [13,] "0"                                         
#> [14,] "0"                                         
#> [15,] "0"                                         
#> [16,] "0"                                         
#>       Environment.Self complementarity Environment.Correction factor
#>  [1,] "FALSE"                          "4"                          
#>  [2,] "FALSE"                          "4"                          
#>  [3,] "FALSE"                          "4"                          
#>  [4,] "FALSE"                          "4"                          
#>  [5,] "FALSE"                          "4"                          
#>  [6,] "FALSE"                          "4"                          
#>  [7,] "FALSE"                          "4"                          
#>  [8,] "FALSE"                          "4"                          
#>  [9,] "FALSE"                          "4"                          
#> [10,] "FALSE"                          "4"                          
#> [11,] "FALSE"                          "4"                          
#> [12,] "FALSE"                          "4"                          
#> [13,] "FALSE"                          "4"                          
#> [14,] "FALSE"                          "4"                          
#> [15,] "FALSE"                          "4"                          
#> [16,] "FALSE"                          "4"                          
#>       Options.Approximative formula Options.Nearest neighbour model
#>  [1,] NA                            "all97"                        
#>  [2,] NA                            "all97"                        
#>  [3,] NA                            "all97"                        
#>  [4,] NA                            "all97"                        
#>  [5,] NA                            "all97"                        
#>  [6,] NA                            "all97"                        
#>  [7,] NA                            "all97"                        
#>  [8,] NA                            "all97"                        
#>  [9,] NA                            "all97"                        
#> [10,] NA                            "all97"                        
#> [11,] NA                            "all97"                        
#> [12,] NA                            "all97"                        
#> [13,] NA                            "all97"                        
#> [14,] NA                            "all97"                        
#> [15,] NA                            "all97"                        
#> [16,] NA                            "all97"                        
#>       Options.GU model Options.Single mismatch model
#>  [1,] NA               "allsanpey"                  
#>  [2,] NA               "allsanpey"                  
#>  [3,] NA               "allsanpey"                  
#>  [4,] NA               "allsanpey"                  
#>  [5,] NA               "allsanpey"                  
#>  [6,] NA               "allsanpey"                  
#>  [7,] NA               "allsanpey"                  
#>  [8,] NA               "allsanpey"                  
#>  [9,] NA               "allsanpey"                  
#> [10,] NA               "allsanpey"                  
#> [11,] NA               "allsanpey"                  
#> [12,] NA               "allsanpey"                  
#> [13,] NA               "allsanpey"                  
#> [14,] NA               "allsanpey"                  
#> [15,] NA               "allsanpey"                  
#> [16,] NA               "allsanpey"                  
#>       Options.Tandem mismatch model Options.Single dangling end model
#>  [1,] "allsanpey"                   "bom00"                          
#>  [2,] "allsanpey"                   "bom00"                          
#>  [3,] "allsanpey"                   "bom00"                          
#>  [4,] "allsanpey"                   "bom00"                          
#>  [5,] "allsanpey"                   "bom00"                          
#>  [6,] "allsanpey"                   "bom00"                          
#>  [7,] "allsanpey"                   "bom00"                          
#>  [8,] "allsanpey"                   "bom00"                          
#>  [9,] "allsanpey"                   "bom00"                          
#> [10,] "allsanpey"                   "bom00"                          
#> [11,] "allsanpey"                   "bom00"                          
#> [12,] "allsanpey"                   "bom00"                          
#> [13,] "allsanpey"                   "bom00"                          
#> [14,] "allsanpey"                   "bom00"                          
#> [15,] "allsanpey"                   "bom00"                          
#> [16,] "allsanpey"                   "bom00"                          
#>       Options.Double dangling end model Options.Long dangling end model
#>  [1,] "sugdna02"                        "sugdna02"                     
#>  [2,] "sugdna02"                        "sugdna02"                     
#>  [3,] "sugdna02"                        "sugdna02"                     
#>  [4,] "sugdna02"                        "sugdna02"                     
#>  [5,] "sugdna02"                        "sugdna02"                     
#>  [6,] "sugdna02"                        "sugdna02"                     
#>  [7,] "sugdna02"                        "sugdna02"                     
#>  [8,] "sugdna02"                        "sugdna02"                     
#>  [9,] "sugdna02"                        "sugdna02"                     
#> [10,] "sugdna02"                        "sugdna02"                     
#> [11,] "sugdna02"                        "sugdna02"                     
#> [12,] "sugdna02"                        "sugdna02"                     
#> [13,] "sugdna02"                        "sugdna02"                     
#> [14,] "sugdna02"                        "sugdna02"                     
#> [15,] "sugdna02"                        "sugdna02"                     
#> [16,] "sugdna02"                        "sugdna02"                     
#>       Options.Internal loop model Options.Single bulge loop model
#>  [1,] "san04"                     "tan04"                        
#>  [2,] "san04"                     "tan04"                        
#>  [3,] "san04"                     "tan04"                        
#>  [4,] "san04"                     "tan04"                        
#>  [5,] "san04"                     "tan04"                        
#>  [6,] "san04"                     "tan04"                        
#>  [7,] "san04"                     "tan04"                        
#>  [8,] "san04"                     "tan04"                        
#>  [9,] "san04"                     "tan04"                        
#> [10,] "san04"                     "tan04"                        
#> [11,] "san04"                     "tan04"                        
#> [12,] "san04"                     "tan04"                        
#> [13,] "san04"                     "tan04"                        
#> [14,] "san04"                     "tan04"                        
#> [15,] "san04"                     "tan04"                        
#> [16,] "san04"                     "tan04"                        
#>       Options.Long bulge loop model Options.CNG repeats model
#>  [1,] "san04"                       NA                       
#>  [2,] "san04"                       NA                       
#>  [3,] "san04"                       NA                       
#>  [4,] "san04"                       NA                       
#>  [5,] "san04"                       NA                       
#>  [6,] "san04"                       NA                       
#>  [7,] "san04"                       NA                       
#>  [8,] "san04"                       NA                       
#>  [9,] "san04"                       NA                       
#> [10,] "san04"                       NA                       
#> [11,] "san04"                       NA                       
#> [12,] "san04"                       NA                       
#> [13,] "san04"                       NA                       
#> [14,] "san04"                       NA                       
#> [15,] "san04"                       NA                       
#> [16,] "san04"                       NA                       
#>       Options.Inosine bases model Options.Hydroxyadenine bases model
#>  [1,] "san05"                     "sug01"                           
#>  [2,] "san05"                     "sug01"                           
#>  [3,] "san05"                     "sug01"                           
#>  [4,] "san05"                     "sug01"                           
#>  [5,] "san05"                     "sug01"                           
#>  [6,] "san05"                     "sug01"                           
#>  [7,] "san05"                     "sug01"                           
#>  [8,] "san05"                     "sug01"                           
#>  [9,] "san05"                     "sug01"                           
#> [10,] "san05"                     "sug01"                           
#> [11,] "san05"                     "sug01"                           
#> [12,] "san05"                     "sug01"                           
#> [13,] "san05"                     "sug01"                           
#> [14,] "san05"                     "sug01"                           
#> [15,] "san05"                     "sug01"                           
#> [16,] "san05"                     "sug01"                           
#>       Options.Azobenzenes model Options.Locked nucleic acids model
#>  [1,] "asa05"                   "owc11"                           
#>  [2,] "asa05"                   "owc11"                           
#>  [3,] "asa05"                   "owc11"                           
#>  [4,] "asa05"                   "owc11"                           
#>  [5,] "asa05"                   "owc11"                           
#>  [6,] "asa05"                   "owc11"                           
#>  [7,] "asa05"                   "owc11"                           
#>  [8,] "asa05"                   "owc11"                           
#>  [9,] "asa05"                   "owc11"                           
#> [10,] "asa05"                   "owc11"                           
#> [11,] "asa05"                   "owc11"                           
#> [12,] "asa05"                   "owc11"                           
#> [13,] "asa05"                   "owc11"                           
#> [14,] "asa05"                   "owc11"                           
#> [15,] "asa05"                   "owc11"                           
#> [16,] "asa05"                   "owc11"                           
#>       Options.Ion correction method Options.Na equivalence correction method
#>  [1,] NA                            "ahs01"                                 
#>  [2,] NA                            "ahs01"                                 
#>  [3,] NA                            "ahs01"                                 
#>  [4,] NA                            "ahs01"                                 
#>  [5,] NA                            "ahs01"                                 
#>  [6,] NA                            "ahs01"                                 
#>  [7,] NA                            "ahs01"                                 
#>  [8,] NA                            "ahs01"                                 
#>  [9,] NA                            "ahs01"                                 
#> [10,] NA                            "ahs01"                                 
#> [11,] NA                            "ahs01"                                 
#> [12,] NA                            "ahs01"                                 
#> [13,] NA                            "ahs01"                                 
#> [14,] NA                            "ahs01"                                 
#> [15,] NA                            "ahs01"                                 
#> [16,] NA                            "ahs01"                                 
#>       Options.DMSO correction method Options.Formamide correction method
#>  [1,] "ahs01"                        "bla96"                            
#>  [2,] "ahs01"                        "bla96"                            
#>  [3,] "ahs01"                        "bla96"                            
#>  [4,] "ahs01"                        "bla96"                            
#>  [5,] "ahs01"                        "bla96"                            
#>  [6,] "ahs01"                        "bla96"                            
#>  [7,] "ahs01"                        "bla96"                            
#>  [8,] "ahs01"                        "bla96"                            
#>  [9,] "ahs01"                        "bla96"                            
#> [10,] "ahs01"                        "bla96"                            
#> [11,] "ahs01"                        "bla96"                            
#> [12,] "ahs01"                        "bla96"                            
#> [13,] "ahs01"                        "bla96"                            
#> [14,] "ahs01"                        "bla96"                            
#> [15,] "ahs01"                        "bla96"                            
#> [16,] "ahs01"                        "bla96"                            
#>       Options.Mode Results.Enthalpy (cal) Results.Entropy (cal)
#>  [1,] "def"        "-47700"               "-138.1"             
#>  [2,] "def"        "-55600"               "-160.3"             
#>  [3,] "def"        "-78800"               "-229.7"             
#>  [4,] "def"        "-83500"               "-227"               
#>  [5,] "def"        "-64600"               "-183.2"             
#>  [6,] "def"        "-81100"               "-222.5"             
#>  [7,] "def"        "-63500"               "-182.5"             
#>  [8,] "def"        "-41200"               "-117.5"             
#>  [9,] "def"        "-50700"               "-147.2"             
#> [10,] "def"        "-113800"              "-323.6"             
#> [11,] "def"        "-62100"               "-179.8"             
#> [12,] "def"        "-46000"               "-124.8"             
#> [13,] "def"        "-40400"               "-109.9"             
#> [14,] "def"        "-63700"               "-181.3"             
#> [15,] "def"        "-90400"               "-249.5"             
#> [16,] "def"        "-80300"               "-218.6"             
#>       Results.Enthalpy (J) Results.Entropy (J) Results.Melting temperature (C)
#>  [1,] "-199386"            "-577.258"          "31.7814953255144"             
#>  [2,] "-232408"            "-670.054"          "38.1103863719918"             
#>  [3,] "-329384"            "-960.146"          "44.5553259145469"             
#>  [4,] "-349030"            "-948.86"           "67.2098590318261"             
#>  [5,] "-270028"            "-765.776"          "47.400072116762"              
#>  [6,] "-338998"            "-930.05"           "63.6040550863501"             
#>  [7,] "-265430"            "-762.85"           "43.0400604037136"             
#>  [8,] "-172216"            "-491.15"           "30.1735038367475"             
#>  [9,] "-211926"            "-615.296"          "33.1415226764116"             
#> [10,] "-475684"            "-1352.648"         "59.6680431282422"             
#> [11,] "-259578"            "-751.564"          "40.2828264688437"             
#> [12,] "-192280"            "-521.664"          "48.2393469411973"             
#> [13,] "-168872"            "-459.382"          "41.9123740666287"             
#> [14,] "-266266"            "-757.834"          "45.9425910944819"             
#> [15,] "-377872"            "-1042.91"          "64.379329012421"              
#> [16,] "-335654"            "-913.748"          "65.7707030297917"             
#>       Message
#>  [1,] NA     
#>  [2,] NA     
#>  [3,] NA     
#>  [4,] NA     
#>  [5,] NA     
#>  [6,] NA     
#>  [7,] NA     
#>  [8,] NA     
#>  [9,] NA     
#> [10,] NA     
#> [11,] NA     
#> [12,] NA     
#> [13,] NA     
#> [14,] NA     
#> [15,] NA     
#> [16,] NA     

seq <- c("GCAUACG", "CAGUAGGUC", "CGCUCGC", "GAGUGGAG", "GACAGGCUG",
         "CAGUACGUC", "GACAUCCUG", "GACCACCUG", "CAGAAUGUC", "GCGUCGC",
         "CGUCCGG", "GACUCUCUG", "CAGCUGGUC", "GACUAGCUG", "CUCUGCUC",
         "GCGUCCG", "GUCCGCG", "CGAUCAC", "GACUACCUG", "GACGAUCUG")

comp.seq <- c("CGUUUGC", "GUCGGCCAG", "GCGUGCG", "CUCUUCUC", "CUGUGCGAC",
              "GUCGGGCAG", "CUGUUGGAC", "CUGGGGGAC", "GUCUGGCAG", "CGCUGCG",
              "GCUGGCC", "CUGAUAGAC", "GUCGUUCAG", "CUGAGCGAC", "GAGUUGAG",
              "CGCUGGC", "CUGGCGC", "GCUUGUG", "CUGAGGGAC", "CUGCCAGAC")

meltingBatch(sequence = seq, comp.seq = comp.seq, nucleic.acid.conc = 0.0004,
             hybridisation.type = "rnarna", Na.conc = 1,
             method.singleMM = "tur06")
#>       Environment.Sequence Environment.Complementary sequence
#>  [1,] "GCAUACG"            "CGUUUGC"                         
#>  [2,] "CAGUAGGUC"          "GUCGGCCAG"                       
#>  [3,] "CGCUCGC"            "GCGUGCG"                         
#>  [4,] "GAGUGGAG"           "CUCUUCUC"                        
#>  [5,] "GACAGGCUG"          "CUGUGCGAC"                       
#>  [6,] "CAGUACGUC"          "GUCGGGCAG"                       
#>  [7,] "GACAUCCUG"          "CUGUUGGAC"                       
#>  [8,] "GACCACCUG"          "CUGGGGGAC"                       
#>  [9,] "CAGAAUGUC"          "GUCUGGCAG"                       
#> [10,] "GCGUCGC"            "CGCUGCG"                         
#> [11,] "CGUCCGG"            "GCUGGCC"                         
#> [12,] "GACUCUCUG"          "CUGAUAGAC"                       
#> [13,] "CAGCUGGUC"          "GUCGUUCAG"                       
#> [14,] "GACUAGCUG"          "CUGAGCGAC"                       
#> [15,] "CUCUGCUC"           "GAGUUGAG"                        
#> [16,] "GCGUCCG"            "CGCUGGC"                         
#> [17,] "GUCCGCG"            "CUGGCGC"                         
#> [18,] "CGAUCAC"            "GCUUGUG"                         
#> [19,] "GACUACCUG"          "CUGAGGGAC"                       
#> [20,] "GACGAUCUG"          "CUGCCAGAC"                       
#>       Environment.Nucleic acid concentration (M) Environment.Hybridization type
#>  [1,] "4e-04"                                    "rnarna"                      
#>  [2,] "4e-04"                                    "rnarna"                      
#>  [3,] "4e-04"                                    "rnarna"                      
#>  [4,] "4e-04"                                    "rnarna"                      
#>  [5,] "4e-04"                                    "rnarna"                      
#>  [6,] "4e-04"                                    "rnarna"                      
#>  [7,] "4e-04"                                    "rnarna"                      
#>  [8,] "4e-04"                                    "rnarna"                      
#>  [9,] "4e-04"                                    "rnarna"                      
#> [10,] "4e-04"                                    "rnarna"                      
#> [11,] "4e-04"                                    "rnarna"                      
#> [12,] "4e-04"                                    "rnarna"                      
#> [13,] "4e-04"                                    "rnarna"                      
#> [14,] "4e-04"                                    "rnarna"                      
#> [15,] "4e-04"                                    "rnarna"                      
#> [16,] "4e-04"                                    "rnarna"                      
#> [17,] "4e-04"                                    "rnarna"                      
#> [18,] "4e-04"                                    "rnarna"                      
#> [19,] "4e-04"                                    "rnarna"                      
#> [20,] "4e-04"                                    "rnarna"                      
#>       Environment.Na concentration (M) Environment.Mg concentration (M)
#>  [1,] "1"                              "0"                             
#>  [2,] "1"                              "0"                             
#>  [3,] "1"                              "0"                             
#>  [4,] "1"                              "0"                             
#>  [5,] "1"                              "0"                             
#>  [6,] "1"                              "0"                             
#>  [7,] "1"                              "0"                             
#>  [8,] "1"                              "0"                             
#>  [9,] "1"                              "0"                             
#> [10,] "1"                              "0"                             
#> [11,] "1"                              "0"                             
#> [12,] "1"                              "0"                             
#> [13,] "1"                              "0"                             
#> [14,] "1"                              "0"                             
#> [15,] "1"                              "0"                             
#> [16,] "1"                              "0"                             
#> [17,] "1"                              "0"                             
#> [18,] "1"                              "0"                             
#> [19,] "1"                              "0"                             
#> [20,] "1"                              "0"                             
#>       Environment.Tris concentration (M) Environment.K concentration (M)
#>  [1,] "0"                                "0"                            
#>  [2,] "0"                                "0"                            
#>  [3,] "0"                                "0"                            
#>  [4,] "0"                                "0"                            
#>  [5,] "0"                                "0"                            
#>  [6,] "0"                                "0"                            
#>  [7,] "0"                                "0"                            
#>  [8,] "0"                                "0"                            
#>  [9,] "0"                                "0"                            
#> [10,] "0"                                "0"                            
#> [11,] "0"                                "0"                            
#> [12,] "0"                                "0"                            
#> [13,] "0"                                "0"                            
#> [14,] "0"                                "0"                            
#> [15,] "0"                                "0"                            
#> [16,] "0"                                "0"                            
#> [17,] "0"                                "0"                            
#> [18,] "0"                                "0"                            
#> [19,] "0"                                "0"                            
#> [20,] "0"                                "0"                            
#>       Environment.dNTP concentration (M) Environment.DMSO concentration (%)
#>  [1,] "0"                                "0"                               
#>  [2,] "0"                                "0"                               
#>  [3,] "0"                                "0"                               
#>  [4,] "0"                                "0"                               
#>  [5,] "0"                                "0"                               
#>  [6,] "0"                                "0"                               
#>  [7,] "0"                                "0"                               
#>  [8,] "0"                                "0"                               
#>  [9,] "0"                                "0"                               
#> [10,] "0"                                "0"                               
#> [11,] "0"                                "0"                               
#> [12,] "0"                                "0"                               
#> [13,] "0"                                "0"                               
#> [14,] "0"                                "0"                               
#> [15,] "0"                                "0"                               
#> [16,] "0"                                "0"                               
#> [17,] "0"                                "0"                               
#> [18,] "0"                                "0"                               
#> [19,] "0"                                "0"                               
#> [20,] "0"                                "0"                               
#>       Environment.Formamide concentration (M or %)
#>  [1,] "0"                                         
#>  [2,] "0"                                         
#>  [3,] "0"                                         
#>  [4,] "0"                                         
#>  [5,] "0"                                         
#>  [6,] "0"                                         
#>  [7,] "0"                                         
#>  [8,] "0"                                         
#>  [9,] "0"                                         
#> [10,] "0"                                         
#> [11,] "0"                                         
#> [12,] "0"                                         
#> [13,] "0"                                         
#> [14,] "0"                                         
#> [15,] "0"                                         
#> [16,] "0"                                         
#> [17,] "0"                                         
#> [18,] "0"                                         
#> [19,] "0"                                         
#> [20,] "0"                                         
#>       Environment.Self complementarity Environment.Correction factor
#>  [1,] "FALSE"                          "4"                          
#>  [2,] "FALSE"                          "4"                          
#>  [3,] "FALSE"                          "4"                          
#>  [4,] "FALSE"                          "4"                          
#>  [5,] "FALSE"                          "4"                          
#>  [6,] "FALSE"                          "4"                          
#>  [7,] "FALSE"                          "4"                          
#>  [8,] "FALSE"                          "4"                          
#>  [9,] "FALSE"                          "4"                          
#> [10,] "FALSE"                          "4"                          
#> [11,] "FALSE"                          "4"                          
#> [12,] "FALSE"                          "4"                          
#> [13,] "FALSE"                          "4"                          
#> [14,] "FALSE"                          "4"                          
#> [15,] "FALSE"                          "4"                          
#> [16,] "FALSE"                          "4"                          
#> [17,] "FALSE"                          "4"                          
#> [18,] "FALSE"                          "4"                          
#> [19,] "FALSE"                          "4"                          
#> [20,] "FALSE"                          "4"                          
#>       Options.Approximative formula Options.Nearest neighbour model
#>  [1,] NA                            "xia98"                        
#>  [2,] NA                            "xia98"                        
#>  [3,] NA                            "xia98"                        
#>  [4,] NA                            "xia98"                        
#>  [5,] NA                            "xia98"                        
#>  [6,] NA                            "xia98"                        
#>  [7,] NA                            "xia98"                        
#>  [8,] NA                            "xia98"                        
#>  [9,] NA                            "xia98"                        
#> [10,] NA                            "xia98"                        
#> [11,] NA                            "xia98"                        
#> [12,] NA                            "xia98"                        
#> [13,] NA                            "xia98"                        
#> [14,] NA                            "xia98"                        
#> [15,] NA                            "xia98"                        
#> [16,] NA                            "xia98"                        
#> [17,] NA                            "xia98"                        
#> [18,] NA                            "xia98"                        
#> [19,] NA                            "xia98"                        
#> [20,] NA                            "xia98"                        
#>       Options.GU model Options.Single mismatch model
#>  [1,] "ser12"          "tur06"                      
#>  [2,] "ser12"          "tur06"                      
#>  [3,] "ser12"          "tur06"                      
#>  [4,] "ser12"          "tur06"                      
#>  [5,] "ser12"          "tur06"                      
#>  [6,] "ser12"          "tur06"                      
#>  [7,] "ser12"          "tur06"                      
#>  [8,] "ser12"          "tur06"                      
#>  [9,] "ser12"          "tur06"                      
#> [10,] "ser12"          "tur06"                      
#> [11,] "ser12"          "tur06"                      
#> [12,] "ser12"          "tur06"                      
#> [13,] "ser12"          "tur06"                      
#> [14,] "ser12"          "tur06"                      
#> [15,] "ser12"          "tur06"                      
#> [16,] "ser12"          "tur06"                      
#> [17,] "ser12"          "tur06"                      
#> [18,] "ser12"          "tur06"                      
#> [19,] "ser12"          "tur06"                      
#> [20,] "ser12"          "tur06"                      
#>       Options.Tandem mismatch model Options.Single dangling end model
#>  [1,] "tur06"                       "ser08"                          
#>  [2,] "tur06"                       "ser08"                          
#>  [3,] "tur06"                       "ser08"                          
#>  [4,] "tur06"                       "ser08"                          
#>  [5,] "tur06"                       "ser08"                          
#>  [6,] "tur06"                       "ser08"                          
#>  [7,] "tur06"                       "ser08"                          
#>  [8,] "tur06"                       "ser08"                          
#>  [9,] "tur06"                       "ser08"                          
#> [10,] "tur06"                       "ser08"                          
#> [11,] "tur06"                       "ser08"                          
#> [12,] "tur06"                       "ser08"                          
#> [13,] "tur06"                       "ser08"                          
#> [14,] "tur06"                       "ser08"                          
#> [15,] "tur06"                       "ser08"                          
#> [16,] "tur06"                       "ser08"                          
#> [17,] "tur06"                       "ser08"                          
#> [18,] "tur06"                       "ser08"                          
#> [19,] "tur06"                       "ser08"                          
#> [20,] "tur06"                       "ser08"                          
#>       Options.Double dangling end model Options.Long dangling end model
#>  [1,] "ser06"                           "sugrna02"                     
#>  [2,] "ser06"                           "sugrna02"                     
#>  [3,] "ser06"                           "sugrna02"                     
#>  [4,] "ser06"                           "sugrna02"                     
#>  [5,] "ser06"                           "sugrna02"                     
#>  [6,] "ser06"                           "sugrna02"                     
#>  [7,] "ser06"                           "sugrna02"                     
#>  [8,] "ser06"                           "sugrna02"                     
#>  [9,] "ser06"                           "sugrna02"                     
#> [10,] "ser06"                           "sugrna02"                     
#> [11,] "ser06"                           "sugrna02"                     
#> [12,] "ser06"                           "sugrna02"                     
#> [13,] "ser06"                           "sugrna02"                     
#> [14,] "ser06"                           "sugrna02"                     
#> [15,] "ser06"                           "sugrna02"                     
#> [16,] "ser06"                           "sugrna02"                     
#> [17,] "ser06"                           "sugrna02"                     
#> [18,] "ser06"                           "sugrna02"                     
#> [19,] "ser06"                           "sugrna02"                     
#> [20,] "ser06"                           "sugrna02"                     
#>       Options.Internal loop model Options.Single bulge loop model
#>  [1,] "tur06"                     "tur06"                        
#>  [2,] "tur06"                     "tur06"                        
#>  [3,] "tur06"                     "tur06"                        
#>  [4,] "tur06"                     "tur06"                        
#>  [5,] "tur06"                     "tur06"                        
#>  [6,] "tur06"                     "tur06"                        
#>  [7,] "tur06"                     "tur06"                        
#>  [8,] "tur06"                     "tur06"                        
#>  [9,] "tur06"                     "tur06"                        
#> [10,] "tur06"                     "tur06"                        
#> [11,] "tur06"                     "tur06"                        
#> [12,] "tur06"                     "tur06"                        
#> [13,] "tur06"                     "tur06"                        
#> [14,] "tur06"                     "tur06"                        
#> [15,] "tur06"                     "tur06"                        
#> [16,] "tur06"                     "tur06"                        
#> [17,] "tur06"                     "tur06"                        
#> [18,] "tur06"                     "tur06"                        
#> [19,] "tur06"                     "tur06"                        
#> [20,] "tur06"                     "tur06"                        
#>       Options.Long bulge loop model Options.CNG repeats model
#>  [1,] "tur06"                       NA                       
#>  [2,] "tur06"                       NA                       
#>  [3,] "tur06"                       NA                       
#>  [4,] "tur06"                       NA                       
#>  [5,] "tur06"                       NA                       
#>  [6,] "tur06"                       NA                       
#>  [7,] "tur06"                       NA                       
#>  [8,] "tur06"                       NA                       
#>  [9,] "tur06"                       NA                       
#> [10,] "tur06"                       NA                       
#> [11,] "tur06"                       NA                       
#> [12,] "tur06"                       NA                       
#> [13,] "tur06"                       NA                       
#> [14,] "tur06"                       NA                       
#> [15,] "tur06"                       NA                       
#> [16,] "tur06"                       NA                       
#> [17,] "tur06"                       NA                       
#> [18,] "tur06"                       NA                       
#> [19,] "tur06"                       NA                       
#> [20,] "tur06"                       NA                       
#>       Options.Inosine bases model Options.Hydroxyadenine bases model
#>  [1,] "zno07"                     NA                                
#>  [2,] "zno07"                     NA                                
#>  [3,] "zno07"                     NA                                
#>  [4,] "zno07"                     NA                                
#>  [5,] "zno07"                     NA                                
#>  [6,] "zno07"                     NA                                
#>  [7,] "zno07"                     NA                                
#>  [8,] "zno07"                     NA                                
#>  [9,] "zno07"                     NA                                
#> [10,] "zno07"                     NA                                
#> [11,] "zno07"                     NA                                
#> [12,] "zno07"                     NA                                
#> [13,] "zno07"                     NA                                
#> [14,] "zno07"                     NA                                
#> [15,] "zno07"                     NA                                
#> [16,] "zno07"                     NA                                
#> [17,] "zno07"                     NA                                
#> [18,] "zno07"                     NA                                
#> [19,] "zno07"                     NA                                
#> [20,] "zno07"                     NA                                
#>       Options.Azobenzenes model Options.Locked nucleic acids model
#>  [1,] NA                        NA                                
#>  [2,] NA                        NA                                
#>  [3,] NA                        NA                                
#>  [4,] NA                        NA                                
#>  [5,] NA                        NA                                
#>  [6,] NA                        NA                                
#>  [7,] NA                        NA                                
#>  [8,] NA                        NA                                
#>  [9,] NA                        NA                                
#> [10,] NA                        NA                                
#> [11,] NA                        NA                                
#> [12,] NA                        NA                                
#> [13,] NA                        NA                                
#> [14,] NA                        NA                                
#> [15,] NA                        NA                                
#> [16,] NA                        NA                                
#> [17,] NA                        NA                                
#> [18,] NA                        NA                                
#> [19,] NA                        NA                                
#> [20,] NA                        NA                                
#>       Options.Ion correction method Options.Na equivalence correction method
#>  [1,] NA                            "ahs01"                                 
#>  [2,] NA                            "ahs01"                                 
#>  [3,] NA                            "ahs01"                                 
#>  [4,] NA                            "ahs01"                                 
#>  [5,] NA                            "ahs01"                                 
#>  [6,] NA                            "ahs01"                                 
#>  [7,] NA                            "ahs01"                                 
#>  [8,] NA                            "ahs01"                                 
#>  [9,] NA                            "ahs01"                                 
#> [10,] NA                            "ahs01"                                 
#> [11,] NA                            "ahs01"                                 
#> [12,] NA                            "ahs01"                                 
#> [13,] NA                            "ahs01"                                 
#> [14,] NA                            "ahs01"                                 
#> [15,] NA                            "ahs01"                                 
#> [16,] NA                            "ahs01"                                 
#> [17,] NA                            "ahs01"                                 
#> [18,] NA                            "ahs01"                                 
#> [19,] NA                            "ahs01"                                 
#> [20,] NA                            "ahs01"                                 
#>       Options.DMSO correction method Options.Formamide correction method
#>  [1,] "ahs01"                        "bla96"                            
#>  [2,] "ahs01"                        "bla96"                            
#>  [3,] "ahs01"                        "bla96"                            
#>  [4,] "ahs01"                        "bla96"                            
#>  [5,] "ahs01"                        "bla96"                            
#>  [6,] "ahs01"                        "bla96"                            
#>  [7,] "ahs01"                        "bla96"                            
#>  [8,] "ahs01"                        "bla96"                            
#>  [9,] "ahs01"                        "bla96"                            
#> [10,] "ahs01"                        "bla96"                            
#> [11,] "ahs01"                        "bla96"                            
#> [12,] "ahs01"                        "bla96"                            
#> [13,] "ahs01"                        "bla96"                            
#> [14,] "ahs01"                        "bla96"                            
#> [15,] "ahs01"                        "bla96"                            
#> [16,] "ahs01"                        "bla96"                            
#> [17,] "ahs01"                        "bla96"                            
#> [18,] "ahs01"                        "bla96"                            
#> [19,] "ahs01"                        "bla96"                            
#> [20,] "ahs01"                        "bla96"                            
#>       Options.Mode Results.Enthalpy (cal) Results.Entropy (cal)
#>  [1,] "def"        "-47650"               "-138.8"             
#>  [2,] "def"        "-71130"               "-200.5"             
#>  [3,] "def"        "-57930"               "-164.1"             
#>  [4,] "def"        "-60570"               "-176.6"             
#>  [5,] "def"        "-79870"               "-219.9"             
#>  [6,] "def"        "-68380"               "-194.5"             
#>  [7,] "def"        "-73880"               "-208.3"             
#>  [8,] "def"        "-78430"               "-218.3"             
#>  [9,] "def"        "-59650"               "-171.5"             
#> [10,] "def"        "-61330"               "-173.8"             
#> [11,] "def"        "-58350"               "-165.4"             
#> [12,] "def"        "-64570"               "-184.7"             
#> [13,] "def"        "-70970"               "-200.6"             
#> [14,] "def"        "-72010"               "-203"               
#> [15,] "def"        "-58820"               "-171"               
#> [16,] "def"        "-59840"               "-169.6"             
#> [17,] "def"        "-59840"               "-169.6"             
#> [18,] "def"        "-50210"               "-148.3"             
#> [19,] "def"        "-70520"               "-198.8"             
#> [20,] "def"        "-69730"               "-198.2"             
#>       Results.Enthalpy (J) Results.Entropy (J) Results.Melting temperature (C)
#>  [1,] "-199177"            "-580.184"          "30.1048299398322"             
#>  [2,] "-297323.4"          "-838.09"           "51.8989532242754"             
#>  [3,] "-242147.4"          "-685.938"          "44.3989325444856"             
#>  [4,] "-253182.6"          "-738.188"          "37.5791954133529"             
#>  [5,] "-333856.6"          "-919.182"          "62.1162425798375"             
#>  [6,] "-285828.4"          "-813.01"           "48.141439592185"              
#>  [7,] "-308818.4"          "-870.694"          "52.845957204839"              
#>  [8,] "-327837.4"          "-912.494"          "58.2977096620104"             
#>  [9,] "-249337"            "-716.87"           "41.08087522322"               
#> [10,] "-256359.4"          "-726.484"          "46.0633145887674"             
#> [11,] "-243903"            "-691.372"          "44.4380466975271"             
#> [12,] "-269902.6"          "-772.046"          "44.8840464672343"             
#> [13,] "-296654.6"          "-838.508"          "51.0196486690585"             
#> [14,] "-301001.8"          "-848.54"           "52.203376709706"              
#> [15,] "-245867.6"          "-714.78"           "37.5268181873443"             
#> [16,] "-250131.2"          "-708.928"          "45.2688421309843"             
#> [17,] "-250131.2"          "-708.928"          "45.2688421309843"             
#> [18,] "-209877.8"          "-619.894"          "28.1788644808993"             
#> [19,] "-294773.6"          "-830.984"          "51.6345164549562"             
#> [20,] "-291471.4"          "-828.476"          "48.8860141674642"             
#>       Message
#>  [1,] NA     
#>  [2,] NA     
#>  [3,] NA     
#>  [4,] NA     
#>  [5,] NA     
#>  [6,] NA     
#>  [7,] NA     
#>  [8,] NA     
#>  [9,] NA     
#> [10,] NA     
#> [11,] NA     
#> [12,] NA     
#> [13,] NA     
#> [14,] NA     
#> [15,] NA     
#> [16,] NA     
#> [17,] NA     
#> [18,] NA     
#> [19,] NA     
#> [20,] NA