melting

R interface to the MELTING 5 software (Le Novère, 2001; Dumousseau et al., 2012) for computation of enthalpy and entropy of the helix-coil transition, and then the melting temperature of a nucleic acid duplex.

melting(sequence, comp.sequence = NULL,
         nucleic.acid.conc,
         hybridisation.type = c("dnadna", "rnarna", "dnarna",
                               "rnadna", "mrnarna", "rnamrna"),
         Na.conc, Mg.conc, Tris.conc, K.conc,
         dNTP.conc, DMSO.conc, formamide.conc,
         size.threshold = 60, self = FALSE, correction.factor,
         method.approx = c("ahs01", "che93", "che93corr",
                           "marschdot", "owe69", "san98",
                           "wetdna91", "wetrna91", "wetdnarna91"),
         method.nn = c("all97", "bre86", "san04", "san96", "sug96",
                      "tan04", "fre86", "xia98", "sug95", "tur06"),
         method.GU = c("tur99"),
         method.singleMM = c("allsanpey", "tur06", "zno07", "zno08"),
         method.tandemMM = c("allsanpey", "tur99"),
         method.single.dangle = c("bom00", "sugdna02", "sugrna02", "ser08"),
         method.double.dangle = c("sugdna02", "sugrna02", "ser05", "ser06"),
         method.long.dangle = c("sugdna02", "sugrna02"),
         method.internal.loop = c("san04", "tur06", "zno07"),
         method.single.bulge.loop = c("tan04", "san04", "ser07" ,"tur06"),
         method.long.bulge.loop = c("san04", "tur06"),
         method.CNG = c("bro05"),
         method.inosine = c("san05", "zno07"),
         method.hydroxyadenine = c("sug01"),
         method.azobenzenes = c("asa05"),
         method.locked = c("mct04"),
         correction.Na = c("ahs01", "kam71", "marschdot",
                           "owc1904", "owc2004", "owc2104", "owc2204",
                           "san96", "san04", "schlif",
                           "tanna06", "tanna07", "wet91"),
         correction.Mg = c("oxcmg08", "tanmg06", "tanmg07"),
         correction.NaMg = c("oxcmix08", "tanmix07"),
         method.Naeq = c("ahs01", "mit96", "pey00"),
         correction.DMSO = c("ahs01", "cul76", "esc80", "mus80"),

Arguments

sequence
Sequence (5' to 3') of one strand of the nucleic acid duplex as a character string.

comp.sequence
Complementary sequence (3' to 5') of the nucleic acid duplex as a character string.

nucleic.acid.conc
Concentration of the nucleic acid strand (mol L<sup>-1</sup>) in excess as a numeric value.

hybridisation.type
The hybridisation type. Either "dnadna", "rnarna", "dnarna", "rnadna", "mrnarna" or "rnamrna" (see Hybridisation type options).

Na.conc
Concentration of Na ions (M) as a numeric value (see Ion and agent concentrations).

Mg.conc
Concentration of Mg ions (M) as a numeric value (see Ion and agent concentrations).

Tris.conc
Concentration of Tris ions (M) as a numeric value (see Ion and agent concentrations).

K.conc
Concentration of K ions (M) as a numeric value (see Ion and agent concentrations).

dNTP.conc
Concentration of dNTP (M) as a numeric value (see Ion and agent concentrations).

DMSO.conc
Concentration of DMSO (%) as a numeric value (see Ion and agent concentrations).

formamide.conc
Concentration of formamide (M or % depending on correction method) as a numeric value (see Ion and agent concentrations).

size.threshold
Sequence length threshold to decide approximative or nearest-neighbour approach for computation. Default is 60.

self
logical. Specifies that sequence is self complementary and complementary sequence is not required (seed Self complementary sequences). Default is FALSE.

correction.factor
Correction factor to be used to modulate the effect of the nucleic acid concentration (nucleic.acid.conc) in the computation of melting temperature (see Correction factor for nucleic acid concentration).

method.approx
Specify the approximative formula to be used for melting temperature calculation for sequences of length greater than size.threshold. Either "ahs01", "che93", "che93corr", "schdot", "owe69", "san98", "wetdna91", "wetrna91" or "wetdnarna91" (see Approximative formulas).

method.nn
Specify the nearest neighbor model to be used for melting temperature calculation for sequences of length lesser than size.threshold. Either "all97", "bre86", "san04", "san96", "sug96", "tan04", "fre86", "xia98", "sug95" or "tur06" (see Nearest neighbor models).

method.GU
Specify the nearest neighbor model to compute the contribution of GU base pairs to the thermodynamic of helix-coil transition. Available method is "tur99" (see GU wobble base pairs effect).

method.singleMM
Specify the nearest neighbor model to compute the contribution of single mismatch to the thermodynamic of helix-coil transition. Either "allsanpey", "tur06", "zno07" or "zno08" (see Single mismatch effect).

method.tandemMM
Specify the nearest neighbor model to compute the contribution of tandem mismatches to the thermodynamic of helix-coil transition. Either "allsanpey" or "tur99" (see Tandem mismatches effect).

method.single.dangle
Specify the nearest neighbor model to compute the contribution of single dangling end to the thermodynamic of helix-coil transition. Either "bom00", "sugdna02", "sugrna02" or "ser08" (see Single dangling end effect).

method.double.dangle
Specify the nearest neighbor model to compute the contribution of double dangling end to the thermodynamic of helix-coil transition. Either "sugdna02", "sugrna02", "ser05" or "ser06" (see Double dangling end effect).

method.long.dangle
Specify the nearest neighbor model to compute the contribution of long dangling end to the thermodynamic of helix-coil transition. Either "sugdna02" or "sugrna02" (see Long dangling end effect).

method.internal.loop
Specify the nearest neighbor model to compute the contribution of internal loop to the thermodynamic of helix-coil transition. Either "san04", "tur06" or "zno07" (see Internal loop effect).

method.single.bulge.loop
Specify the nearest neighbor model to compute the contribution of single bulge loop to the thermodynamic of helix-coil transition. Either "san04", "tan04", "ser07" or "tur06" (see Single bulge loop effect).

method.long.bulge.loop
Specify the nearest neighbor model to compute the contribution of long bulge loop to the thermodynamic of helix-coil transition. Either "san04" or "tur06" (see Long bulge loop effect).

method.CNG
Specify the nearest neighbor model to compute the contribution of CNG repeats to the thermodynamic of helix-coil transition. Available method is "bro05" (see CNG repeats effect).

method.inosine
Specify the pecific nearest neighbor model to compute the contribution of inosine bases (I) to the thermodynamic of helix-coil transition. Either "san05" or "zno07" (see Inosine bases effect).

method.hydroxyadenine
Specify the nearest neighbor model to compute the contribution of hydroxyadenine bases (A*) to the thermodynamic of helix-coil transition. Available method is "sug01" (see Hydroxyadenine bases effect).

method.azobenzenes
Specify the nearest neighbor model to compute the contribution of azobenzenes (X_T for trans azobenzenes and X_C for cis azobenzenes) to the thermodynamic of helix-coil transition. Available method is "asa05" (see Azobenzenes effect).

method.locked
Specify the nearest neighbor model to compute the contribution of locked nucleic acids (AL, GL, TL and CL) to the thermodynamic of helix-coil transition. Available method is "mct04" (see Locked nucleic acids effect).

correction.Na
Specify the correction method for Na ions. Either "ahs01", "kam71", "owc1904", "owc2004", "owc2104", "owc2204", "san96", "san04", "schlif", "tanna06", "wetdna91", "tanna07", "wetrna91" or "wetdnarna91" (see Sodium corrections).

correction.Mg
Specify the correction method for Mg ions. Either "owcmg08", "tanmg06" or "tanmg07" (see Magnesium corrections).

correction.NaMg
Specify the correction method for mixed Na and Mg ions. Either "owcmix08", "tanmix07" or "tanmix07" (see Mixed Sodium and Magnesium corrections).

method.Naeq
Specify the ion correction which gives a sodium equivalent concentration if other cations are present. Either "ahs01", "mit96" or "pey00" (see Sodium equivalent concentration methods).

correction.DMSO
Specify the correction method for DMSO. Specify the correction method for DMSO. Either "ahs01", "mus81", "cul76" or "esc80" (see DMSO corrections).

correction.formamide
Specify the correction method for formamide. Specify the correction method for formamide Either "bla96" or "lincorr" (see Formamide corrections).

Mandatory arguments

The following are the arguments which are mandatory for computation.

sequence

comp.sequence: Mandatory if there are mismatches, inosine(s) or hydroxyadenine(s) between the two strands. If not specified, it is computed as the complement of sequence. Self-complementarity in sequence is detected even though there may be (are) dangling end(s) and comp.sequence is computed (see Self complementary sequences).

nucleic.acid.conc

Na.conc, Mg.conc, Tris.conc, K.conc: At least one cation (Na, Mg, Tris, K) concentration is mandatory, the other agents(dNTP, DMSO, formamide) are optional.

hybridisation.type

Hybridisation type options

The details of the possible options for hybridisation type specified in the argument hybridisation.type are as follows:

Option	Sequence	Complementary sequence

dnadna	DNA	DNA

rnarna	RNA	RNA

dnarna	DNA	RNA

rnadna	RNA	DNA

mrnarna	2-o-methyl RNA	RNA

This parameter determines the nature of the sequences in the arguments sequence and comp.sequence.

Ion and agent concentrations

These values are used for different correction functions which approximately adjusts for effects of these ions(Na, Mg, Tris, K) and/or agents(dNTP, DMSO, formamide) on on thermodynamic stability of nucleic acid duplexes. Their concentration limits depends on the correction method used. All the concentrations must be in M, except for the DMSO (%) and formamide (% or M depending on the correction method). Note that Tris+ concentration is about half of the total tris buffer concentration.

Self complementary sequences

Self complementarity for perfect matching sequences or sequences with dangling ends is detected automatically. However it can be specified by the argument self.

Correction factor for nucleic acid concentration

For self complementary sequences (Auto detected or specified in self) it is 1. Otherwise it is 4 if the both strands are present in equivalent amount and 1 if one strand is in excess.

Approximative estimation formulas

The calculation is increasingly incorrect when the length of the duplex decreases. Moreover, it does not take into account nucleic acid concentration.

Formula	Type	Limits.Remarks	Reference

ahs01	DNA	No mismatch	von Ahsen et al., 2001

che93	DNA	No mismatch; Na=0, Mg=0.0015,	Marmur and Doty, 1962

Tris=0.01, K=0.05	che93corr	DNA

No mismatch; Na=0, Mg=0.0015,	Marmur and Doty, 1962		Tris=0.01, K=0.05

marschdot	DNA	No mismatch	Wetmur, 1991; Marmur and

Doty, 1962; Chester and

Marshak, 1993; Schildkraut

and Lifson, 1965; Wahl et			al., 1987; Britten et al.,

1974; Hall et al., 1980	owe69

DNA	No mismatch	Owen et al., 1969;

Frank-Kamenetskii, 1971;

Blake, 1996; Blake and			Delcourt, 1998

san98	DNA	No mismatch	SantaLucia, 1998; von Ahsen

et al., 2001	wetdna91*

DNA		Wetmur, 1991	wetrna91*

RNA		Wetmur, 1991	Formula

Nearest neighbor models

Model	Type	Limits.Remarks	Reference

all97*	DNA	Allawi and SantaLucia, 1997	tur06*

2'-O-MeRNA/	A sodium correction	Kierzek et al., 2006

RNA	(san04) is		automatically applied to

convert the entropy (Na =		0.1M) into the entropy (Na =

1M).	bre86	DNA

Breslauer et al., 1986	san04	DNA	SantaLucia and Hicks, 2004

san96	DNA	SantaLucia et al., 1996	sug96

DNA	Sugimoto et al., 1996	tan04	DNA

Tanaka et al., 2004	fre86	RNA	Freier et al., 1986

xia98*	RNA	Xia et al., 1998	sug95*

DNA/	SantaLucia et al., 1996	Model	Type

GU wobble base pairs effect

Model	Type	Limits.Remarks	Reference

Single mismatch effect

Model	Type	Limits.Remarks	Reference

allsanpey*	DNA		Allawi and SantaLucia, 1997;

Allawi and SantaLucia, 1998;

Allawi and SantaLucia, 1998;

Allawi and SantaLucia,			1998; Peyret et al., 1999

tur06	RNA		Lu et al., 2006

zno07*	RNA		Davis and Znosko, 2007

zno08	RNA	At least one adjacent GU base	Davis and Znosko, 2008

Tandem mismatches effect

Model	Type	Limits.Remarks	Reference

allsanpey*	DNA	Only GT mismatches and TA/TG	Allawi and SantaLucia, 1997;

mismatches.	Allawi and SantaLucia, 1998;

Allawi and SantaLucia, 1998;		Allawi and SantaLucia,

1998; Peyret et al., 1999	tur99*	RNA	No adjacent GU or UG base

Mathews et al., 1999	Model	Type	Limits.Remarks

Tandem mismatches are not taken into account by the approximative mode. Note that not all the mismatched Crick's pairs have been investigated.

Single dangling end effect

Model	Type	Limits.Remarks	Reference

bom00*	DNA		Bommarito et al., 2000

sugdna02	DNA	Only terminal poly A self	Ohmichi et al., 2002

complementary sequences.		sugrna02

RNA	Only terminal poly A self	Ohmichi et al., 2002

complementary sequences.		ser08*	RNA

Only 3' UA, GU and UG	Miller et al., 2008		terminal base pairs only 5'

UG and GU terminal base

Single dangling ends are not taken into account by the approximative mode.

Double dangling end effect

Model	Type	Limits.Remarks	Reference

sugdna02*	DNA	Only terminal poly A self	Ohmichi et al., 2002

complementary sequences.		sugrna02

RNA	Only terminal poly A self	Ohmichi et al., 2002

complementary sequences.		ser05	RNA

Depends on the available	O'Toole et al., 2005		thermodynamic parameters for

single dangling end.

Double dangling ends are not taken into account by the approximative mode.

Long dangling end effect

Model	Type	Limits.Remarks	Reference

sugdna02*	DNA	Only terminal poly A self	Ohmichi et al., 2002

complementary sequences.		sugrna02*

RNA	Only terminal poly A self	Ohmichi et al., 2002	Model

Long dangling ends are not taken into account by the approximative mode.

Internal loop effect

Model	Type	Limits.Remarks	Reference

san04*	DNA	Missing asymmetry penalty.	SantaLucia and Hicks, 2004

Not tested with experimental

results.		tur06	RNA

Not tested with experimental	Lu et al., 2006		results.

Model	Type	Limits.Remarks

Internal loops are not taken into account by the approximative mode.

Single bulge loop effect

Model	Type	Limits.Remarks	Reference

tan04*	DNA		Tan and Chen, 2007

san04	DNA	Missing closing AT penalty.	SantaLucia and Hicks, 2004

ser07	RNA	Less reliable results. Some	Blose et al., 2007

missing parameters.		Model

Internal loops are not taken into account by the approximative mode.

Long bulge loop effect

Model	Type	Limits.Remarks	Reference

san04*	DNA	Missing closing AT penalty.	SantaLucia and Hicks, 2004

tur06*	RNA	Not tested with experimental	Lu et al., 2006

Long bulge loops are not taken into account by the approximative mode.

CNG repeats effect

Model	Type	Limits.Remarks	Reference

bro05*	RNA	Self complementary sequences.	Broda et al., 2005

CNG repeats are not taken into account by the approximative mode. The contribution of CNG repeats to the thermodynamic of helix-coil transition can be computed only for 2 to 7 CNG repeats. N represents a single mismatch of type N/N.

Inosine bases effect

Model	Type	Limits.Remarks	Reference

san05*	DNA	Missing parameters for tandem	Watkins and SantaLucia, 2005

base pairs containing inosine

bases.		Model	Type

Hydroxyadenine bases effect

Model	Type	Limits.Remarks	Reference

sug01*	DNA	Only 5' GA*C 3'and 5' TA*A 3'	Kawakami et al., 2001

Hydroxyadenine bases (A*) are not taken into account by the approximative mode.

Azobenzenes effect effect

Model	Type	Limits.Remarks	Reference

asa05*	DNA	Less reliable results when	Asanuma et al., 2005

the number of cis azobenzene		Model

Azobenzenes (X_T for trans azobenzenes and X_C for cis azobenzenes) are not taken into account by the approximative mode.

Locked nucleic acids effect

Model	Type	Limits.Remarks	Reference

Locked nucleic acids (AL, GL, TL and CL) are not taken into account by the approximative mode.

Sodium corrections

Correcion	Type	Limits.Remarks	Reference

ahs01	DNA	Na>0.	von Ahsen et al., 2001

schlif	DNA	Na>=0.07; Na<=0.12.	Schildkraut and Lifson, 1965

tanna06	DNA	Na>=0.001; Na<=1.	Tan and Chen, 2006

tanna07*	RNA	Na>=0.003; Na<=1.	Tan and Chen, 2007

or

2'-O-MeRNA/RNA		wet91	RNA,

Na>0.	Wetmur, 1991		DNA

and

RNA/DNA		kam71

DNA	Na>0; Na>=0.069; Na<=1.02.	Frank-Kamenetskii, 1971	marschdot

DNA	Na>=0.069; Na<=1.02.	Marmur and Doty, 1962; Blake

and Delcourt, 1998	owc1904	DNA	Na>0.

Owczarzy et al., 2004	owc2004	DNA	Na>0.

Owczarzy et al., 2004	owc2104	DNA	Na>0.

Owczarzy et al., 2004	owc2204*	DNA	Na>0.

Owczarzy et al., 2004	san96	DNA	Na>=0.1.

SantaLucia et al., 1996	san04	DNA	Na>=0.05; Na<=1.1;

SantaLucia and Hicks, 2004;		Oligonucleotides inferior to	SantaLucia, 1998

Magnesium corrections

Correcion	Type	Limits.Remarks	Reference

owcmg08*	DNA	Mg>=0.0005; Mg<=0.6.	Owczarzy et al., 2008

tanmg06	DNA	Mg>=0.0001; Mg<=1; Oligomer	Tan and Chen, 2006

length superior to 6 base		pairs.

Mixed Sodium and Magnesium corrections

Correcion	Type	Limits.Remarks	Reference

owcmix08*	DNA	Mg>=0.0005; Mg<=0.6;	Owczarzy et al., 2008

Na+K+Tris/2>0.	tanmix07	DNA

Mg>=0.1; Mg<=0.3;	Tan and Chen, 2007		and

Na+K+Tris/2>=0.1;	Correcion	Type	Limits.Remarks

Sodium equivalent concentration methods

Correcion	Type	Limits.Remarks	Reference

ahs01*	DNA	von Ahsen et al., 2001	mit96

DNA	Mitsuhashi, 1996	Correcion	Type

DMSO corrections

Correcion	Type	Limits.Remarks	Reference

ahs01	DNA	Not tested with experimental	von Ahsen et al., 2001

results.	cul76	DNA

Not tested with experimental	Cullen and Bick, 1976		results.

esc80	DNA	Not tested with experimental	Escara and Hutton, 1980

results.	mus80	DNA

Not tested with experimental	Musielski et al., 1981	Correcion	Type

Formamide corrections

Correcion	Type	Limits.Remarks	Reference

bla96	DNA	With formamide concentration	Blake, 1996

in mol/L.		lincorr

DNA	With a formamide volume.	McConaughy et al., 1969;

Record, 1967; Casey and	Correcion	Type

References

Marmur J and Doty P (1962). “Determination of the base composition of deoxyribonucleic acid from its thermal denaturation temperature.” Journal of Molecular Biology, 5(1), pp. 109--118. doi: 10.1016/S0022-2836(62)80066-7 .

Schildkraut C and Lifson S (1965). “Dependence of the melting temperature of DNA on salt concentration.” Biopolymers, 3(2), pp. 195--208. doi: 10.1002/bip.360030207 .

Record MT (1967). “Electrostatic effects on polynucleotide transitions. I. Behavior at neutral pH.” Biopolymers, 5(10), pp. 975--992. doi: 10.1002/bip.1967.360051010 .

McConaughy BL, Laird C and McCarthy BJ (1969). “Nucleic acid reassociation in formamide.” Biochemistry, 8(8), pp. 3289--3295. doi: 10.1021/bi00836a024 .

Owen R, Hill L and Lapage S (1969). “Determination of DNA base compositions from melting profiles in dilute buffers.” Biopolymers, 7(4), pp. 503--516. doi: 10.1002/bip.1969.360070408 .

Frank-Kamenetskii MD (1971). “Simplification of the empirical relationship between melting temperature of DNA, its GC content and concentration of sodium ions in solution.” Biopolymers, 10(12), pp. 2623--2624. doi: 10.1002/bip.360101223 .

Britten RJ, Graham DE and Neufeld BR (1974). “Analysis of repeating DNA sequences by reassociation.” Methods in Enzymology, 29, pp. 363--418. doi: 10.1016/0076-6879(74)29033-5 .

Cullen BR and Bick MD (1976). “Thermal denaturation of DNA from bromodeoxyuridine substituted cells.” Nucleic Acids Research, 3(1), pp. 49--62. doi: 10.1093/nar/3.1.49 .

Hutton JR (1977). “Renaturation kinetics and thermal stability of DNA in aqueous solutions of formamide and urea.” Nucleic Acids Research, 4(10), pp. 3537--3555. doi: 10.1093/nar/4.10.3537 .

Casey J and Davidson N (1977). “Rates of formation and thermal stabilities of RNA:DNA and DNA:DNA duplexes at high concentrations of formamide.” Nucleic Acids Research, 4(5), pp. 1539--1552. doi: 10.1093/nar/4.5.1539 .

Hall TJ, Grula JW, Davidson EH and Britten RJ (1980). “Evolution of sea urchin non-repetitive DNA.” Journal of Molecular Evolution, 16(2), pp. 95--110. doi: 10.1007/BF01731580 .

Escara JF and Hutton JR (1980). “Thermal stability and renaturation of DNA in dimethyl sulfoxide solutions: Acceleration of the renaturation rate.” Biopolymers, 19(7), pp. 1315--1327. doi: 10.1002/bip.1980.360190708 .

Musielski H, Mann W, Laue R and Michel S (1981). “Influence of dimethylsulfoxide on transcription by bacteriophage T3-induced RNA polymerase.” Zeitschrift fur allgemeine Mikrobiologie, 21(6), pp. 447--456. doi: 10.1002/jobm.19810210606 .

Freier SM, Kierzek R, Jaeger JA, Sugimoto N, Caruthers MH, Neilson T and Turner DH (1986). “Improved free-energy parameters for predictions of RNA duplex stability.” Proceedings of the National Academy of Sciences, 83(24), pp. 9373. http://www.pnas.org/content/83/24/9373.abstract.

Breslauer KJ, Frank R, Blocker H and Marky LA (1986). “Predicting DNA duplex stability from the base sequence.” Proceedings of the National Academy of Sciences, 83(11), pp. 3746. http://www.pnas.org/content/83/11/3746.abstract.

Wahl GM, Barger SL and Kimmel AR (1987). “Molecular hybridization of immobilized nucleic acids: Theoretical concepts and practical considerations.” Methods in Enzymology, 152, pp. 399--407. doi: 10.1016/0076-6879(87)52046-8 .

Wetmur JG (1991). “DNA probes: Applications of the principles of nucleic acid hybridization.” Critical Reviews in Biochemistry and Molecular Biology, 26(3-4), pp. 227--259. doi: 10.3109/10409239109114069 .

Chester N and Marshak D (1993). “Dimethyl sulfoxide-mediated primer Tm reduction: A method for analyzing the role of renaturation temperature in the polymerase chain reaction.” Analytical Biochemistry, 209(2), pp. 284--290. doi: 10.1006/abio.1993.1121 .

Sugimoto N, Katoh M, Nakano S, Ohmichi T and Sasaki M (1994). “RNA/DNA hybrid duplexes with identical nearest-neighbor base-pairs have identical stability.” FEBS Letters, 354(1), pp. 74--78. doi: 10.1016/0014-5793(94)01098-6 .

Sugimoto N, Nakano S, Katoh M, Matsumura A, Nakamuta H, Ohmichi T, Yoneyama M and Sasaki M (1995). “Thermodynamic parameters to predict stability of RNA/DNA hybrid duplexes.” Biochemistry, 34(35), pp. 11211--11216. doi: 10.1021/bi00035a029 .

SantaLucia J, Allawi HT and Seneviratne PA (1996). “Improved nearest-neighbor parameters for predicting DNA duplex stability.” Biochemistry, 35(11), pp. 3555--3562. doi: 10.1021/bi951907q .

Sugimoto N, Nakano S, Yoneyama M and Honda K (1996). “Improved thermodynamic parameters and helix initiation factor to predict stability of DNA duplexes.” Nucleic Acids Research, 24(22), pp. 4501--4505. doi: 10.1093/nar/24.22.4501 .

Blake RD and Delcourt SG (1996). “Thermodynamic effects of formamide on DNA stability.” Nucleic Acids Research, 24(11), pp. 2095--2103. doi: 10.1093/nar/24.11.2095 .

Blake RD (1996). “Denaturation of DNA.” In Meyers RA (ed.), Encyclopedia of molecular biology and molecular medicine, volume 2, pp. 1--19. VCH Verlagsgesellschaft, Weinheim, Germany.

Mitsuhashi M (1996). “Technical report: Part 1. Basic requirements for designing optimal oligonucleotide probe sequences.” Journal of Clinical Laboratory Analysis, 10(5), pp. 277--284. doi: 10.1002/(sici)1098-2825(1996)10:5<277::aid-jcla8>3.0.co;2-5 .

Allawi HT and SantaLucia J (1997). “Thermodynamics and NMR of internal G-T mismatches in dna.” Biochemistry, 36(34), pp. 10581--10594. doi: 10.1021/bi962590c .

SantaLucia J (1998). “A unified view of polymer, dumbbell, and oligonucleotide DNA nearest-neighbor thermodynamics.” Proceedings of the National Academy of Sciences, 95(4), pp. 1460. http://www.pnas.org/content/95/4/1460.abstract.

Xia T, SantaLucia J, Burkard ME, Kierzek R, Schroeder SJ, Jiao X, Cox C and Turner DH (1998). “Thermodynamic parameters for an expanded nearest-neighbor model for formation of RNA duplexes with Watson-Crick base pairs.” Biochemistry, 37(42), pp. 14719--14735. doi: 10.1021/bi9809425 .

Allawi HT and SantaLucia J (1998). “Thermodynamics of internal C-T mismatches in DNA.” Nucleic Acids Research, 26(11), pp. 2694--2701. doi: 10.1093/nar/26.11.2694 .

Blake RD and Delcourt SG (1998). “Thermal stability of DNA.” Nucleic Acids Research, 26(14), pp. 3323--3332. doi: 10.1093/nar/26.14.3323 .

Allawi HT and SantaLucia J (1998). “Nearest neighbor thermodynamic parameters for internal G-A mismatches in DNA.” Biochemistry, 37(8), pp. 2170--2179. doi: 10.1021/bi9724873 .

Allawi HT and SantaLucia J (1998). “Nearest-neighbor thermodynamics of internal A-C mismatches in dna: sequence dependence and pH effects.” Biochemistry, 37(26), pp. 9435--9444. doi: 10.1021/bi9803729 .

Mathews DH, Sabina J, Zuker M and Turner DH (1999). “Expanded sequence dependence of thermodynamic parameters improves prediction of RNA secondary structure.” Journal of Molecular Biology, 288(5), pp. 911--940. doi: 10.1006/jmbi.1999.2700 .

Peyret N, Seneviratne PA, Allawi HT and SantaLucia J (1999). “Nearest-Neighbor Thermodynamics and NMR of DNA Sequences with Internal A-A, C-C, G-G, and T-T Mismatches.” Biochemistry, 38(12), pp. 3468--3477. doi: 10.1021/bi9825091 .

Peyret N (2000). Prediction of nucleic acid hybridization: Parameters and algorithms.. PhD thesis, Wayne State University, Detroit, MI.

Bommarito S, Peyret N and SantaLucia J (2000). “Thermodynamic parameters for DNA sequences with dangling ends.” Nucleic Acids Research, 28(9), pp. 1929--1934. doi: 10.1093/nar/28.9.1929 .

Kawakami J, Kamiya H, Yasuda K, Fujiki H, Kasai H and Sugimoto N (2001). “Thermodynamic stability of base pairs between 2-hydroxyadenine and incoming nucleotides as a determinant of nucleotide incorporation specificity during replication.” Nucleic Acids Research, 29(16), pp. 3289--3296. doi: 10.1093/nar/29.16.3289 .

von Ahsen N, Wittwer CT and Schutz E (2001). “Oligonucleotide melting temperatures under PCR conditions: Nearest-neighbor corrections for Mg2+, deoxynucleotide triphosphate, and dimethyl sulfoxide concentrations with comparison to alternative empirical formulas.” Clinical Chemistry, 47(11), pp. 1956--1961. http://clinchem.aaccjnls.org/content/47/11/1956.

Le Novere N (2001). “MELTING, computing the melting temperature of nucleic acid duplex.” Bioinformatics, 17(12), pp. 1226--1227. doi: 10.1093/bioinformatics/17.12.1226 .

Ohmichi T, Nakano S, Miyoshi D and Sugimoto N (2002). “Long RNA dangling end has large energetic contribution to duplex stability.” Journal of the American Chemical Society, 124(35), pp. 10367--10372. doi: 10.1021/ja0255406 .

SantaLucia J and Hicks D (2004). “The thermodynamics of DNA structural motifs.” Annual Review of Biophysics and Biomolecular Structure, 33(1), pp. 415--440. doi: 10.1146/annurev.biophys.32.110601.141800 .

Tanaka F, Kameda A, Yamamoto M and Ohuchi A (2004). “Thermodynamic parameters based on a nearest-neighbor model for DNA sequences with a single-bulge loop.” Biochemistry, 43(22), pp. 7143--7150. doi: 10.1021/bi036188r .

McTigue PM, Peterson RJ and Kahn JD (2004). “Sequence-dependent thermodynamic parameters for locked nucleic acid (LNA)-DNA duplex formation.” Biochemistry, 43(18), pp. 5388--5405. doi: 10.1021/bi035976d .

Owczarzy R, You Y, Moreira BG, Manthey JA, Huang L, Behlke MA and Walder JA (2004). “Effects of sodium ions on DNA duplex oligomers: Improved predictions of melting temperatures.” Biochemistry, 43(12), pp. 3537--3554. doi: 10.1021/bi034621r .

Broda M, Kierzek E, Gdaniec Z, Kulinski T and Kierzek R (2005). “Thermodynamic stability of RNA structures formed by CNG trinucleotide repeats. Implication for prediction of RNA structure.” Biochemistry, 44(32), pp. 10873--10882. doi: 10.1021/bi0502339 .

Watkins NE and SantaLucia J (2005). “Nearest-neighbor thermodynamics of deoxyinosine pairs in DNA duplexes.” Nucleic Acids Research, 33(19), pp. 6258--6267. doi: 10.1093/nar/gki918 .

Asanuma H, Matsunaga D and Komiyama M (2005). “Clear-cut photo-regulation of the formation and dissociation of the DNA duplex by modified oligonucleotide involving multiple azobenzenes.” Nucleic Acids Symposium Series, pp. 35--36. doi: 10.1093/nass/49.1.35 .

O'Toole AS, Miller S and Serra MJ (2005). “Stability of 3' double nucleotide overhangs that model the 3' ends of siRNA.” RNA, 11(4), pp. 512--516. doi: 10.1261/rna.7254905 .

Lu ZJ, Turner DH and Mathews DH (2006). “A set of nearest neighbor parameters for predicting the enthalpy change of RNA secondary structure formation.” Nucleic Acids Research, 34(17), pp. 4912--4924. doi: 10.1093/nar/gkl472 .

Kierzek E, Mathews DH, Ciesielska A, Turner DH and Kierzek R (2006). “Nearest neighbor parameters for Watson-Crick complementary heteroduplexes formed between 2'-O-methyl RNA and RNA oligonucleotides.” Nucleic Acids Research, 34(13), pp. 3609--3614. doi: 10.1093/nar/gkl232 .

Tan Z and Chen S (2006). “Nucleic acid helix stability: Effects of salt concentration, cation valence and size, and chain length.” Biophysical Journal, 90(4), pp. 1175--1190. doi: 10.1529/biophysj.105.070904 .

O'Toole AS, Miller S, Haines N, Zink MC and Serra MJ (2006). “Comprehensive thermodynamic analysis of 3' double-nucleotide overhangs neighboring Watson-Crick terminal base pairs.” Nucleic Acids Research, 34(11), pp. 3338--3344. doi: 10.1093/nar/gkl428 .

Tan Z and Chen S (2007). “RNA helix stability in mixed Na(+)/Mg(2+) solution.” Biophysical Journal, 92(10), pp. 3615--3632. doi: 10.1529/biophysj.106.100388 .

Wright DJ, Rice JL, Yanker DM and Znosko BM (2007). “Nearest neighbor parameters for inosine-uridine pairs in RNA duplexes.” Biochemistry, 46(15), pp. 4625--4634. doi: 10.1021/bi0616910 .

Davis AR and Znosko BM (2007). “Thermodynamic characterization of single mismatches found in naturally occurring RNA.” Biochemistry, 46(46), pp. 13425--13436. doi: 10.1021/bi701311c .

Blose JM, Manni ML, Klapec KA, Stranger-Jones Y, Zyra AC, Sim V, Griffith CA, Long JD and Serra MJ (2007). “Non-nearest-neighbor dependence of stability for RNA bulge loops based on the complete set of group i single nucleotide bulge loops.” Biochemistry, 46(51), pp. 15123--15135. doi: 10.1021/bi700736f .

Badhwar J, Karri S, Cass CK, Wunderlich EL and Znosko BM (2007). “Thermodynamic characterization of RNA duplexes containing naturally occurring 1 * 2 nucleotide internal loops.” Biochemistry, 46(50), pp. 14715--14724. doi: 10.1021/bi701024w .

Davis AR and Znosko BM (2008). “Thermodynamic characterization of naturally occurring RNA single mismatches with G-U nearest neighbors.” Biochemistry, 47(38), pp. 10178--10187. doi: 10.1021/bi800471z .

Miller S, Jones LE, Giovannitti K, Piper D and Serra MJ (2008). “Thermodynamic analysis of 5' and 3' single- and 3' double-nucleotide overhangs neighboring wobble terminal base pairs.” Nucleic Acids Research, 36(17), pp. 5652--5659. doi: 10.1093/nar/gkn525 .

Owczarzy R, Moreira BG, You Y, Behlke MA and Walder JA (2008). “Predicting stability of DNA duplexes in solutions containing magnesium and monovalent cations.” Biochemistry, 47(19), pp. 5336--5353. doi: 10.1021/bi702363u .

Dumousseau M, Rodriguez N, Juty N and Le Novere N (2012). “MELTING, a flexible platform to predict the melting temperatures of nucleic acids.” BMC Bioinformatics, 13, pp. 101. doi: 10.1186/1471-2105-13-101 .

Examples