Compute melting temperature of a nucleic acid duplex

Compute the enthalpy and entropy of helix-coil transition, and then the melting temperature of a nucleic acid duplex with the MELTING 5 software (Le Novère, 2001; Dumousseau et al., 2012).

melting(sequence, comp.sequence = NULL,
        nucleic.acid.conc,
        hybridisation.type = c("dnadna", "rnarna", "dnarna",
                               "rnadna", "mrnarna", "rnamrna"),
        Na.conc, Mg.conc, Tris.conc, K.conc,
        dNTP.conc, DMSO.conc, formamide.conc,
        size.threshold = 60, force.self = FALSE, correction.factor,
        method.approx = c("ahs01", "che93", "che93corr",
                          "schdot", "owe69", "san98",
                          "wetdna91", "wetrna91", "wetdnarna91"),
        method.nn = c("all97", "bre86", "san04", "san96", "sug96",
                      "tan04", "fre86", "xia98", "sug95", "tur06"),
        method.GU = c("tur99", "ser12"),
        method.singleMM = c("allsanpey", "tur06", "zno07", "zno08", "wat11"),
        method.tandemMM = c("allsanpey", "tur99"),
        method.single.dangle = c("bom00", "sugdna02", "sugrna02", "ser08"),
        method.double.dangle = c("sugdna02", "sugrna02", "ser05", "ser06"),
        method.long.dangle = c("sugdna02", "sugrna02"),
        method.internal.loop = c("san04", "tur06", "zno07"),
        method.single.bulge.loop = c("tan04", "san04", "ser07" ,"tur06"),
        method.long.bulge.loop = c("san04", "tur06"),
        method.CNG = c("bro05"),
        method.inosine = c("san05", "zno07"),
        method.hydroxyadenine = c("sug01"),
        method.azobenzenes = c("asa05"),
        method.locked = c("mct04"),
        correction.ion = c("ahs01", "kam71", "marschdot",
                           "owc1904", "owc2004", "owc2104",
                           "owc2204", "san96", "san04", "schlif",
                           "tanna06", "tanna07", "wet91",
                           "owcmg08", "tanmg06", "tanmg07",
                           "owcmix08", "tanmix07"),
        method.Naeq = c("ahs01", "mit96", "pey00"),
        correction.DMSO = c("ahs01", "cul76", "esc80", "mus81"),
        correction.formamide = c("bla96", "lincorr"))

Arguments

sequence
Sequence (5' to 3') of one strand of the nucleic acid duplex as a character string (Note: Uridine and thymidine are not considered as identical).

comp.sequence
Complementary sequence (3' to 5') of the nucleic acid duplex as a character string.

nucleic.acid.conc
Concentration of the nucleic acid strand (M or mol L-1) in excess as a numeric value.

hybridisation.type
The hybridisation type. Either "dnadna", "rnarna", "dnarna", "rnadna", "mrnarna" or "rnamrna" (see Hybridisation type options).

Na.conc
Concentration of Na ions (M) as a positive numeric value (see Ion and agent concentrations).

Mg.conc
Concentration of Mg ions (M) as a positive numeric value (see Ion and agent concentrations).

Tris.conc
Concentration of Tris ions (M) as a positive numeric value (see Ion and agent concentrations).

K.conc
Concentration of K ions (M) as a positive numeric value (see Ion and agent concentrations).

dNTP.conc
Concentration of dNTP (M) as a positive numeric value (see Ion and agent concentrations).

DMSO.conc
Concentration of DMSO (%) as a positive numeric value (see Ion and agent concentrations).

formamide.conc
Concentration of formamide (M or % depending on correction method) as a positive numeric value (see Ion and agent concentrations).

size.threshold
Sequence length threshold to decide approximative or nearest-neighbour approach for computation. Default is 60.

force.self
logical. Enforces that sequence is self complementary and complementary sequence is not required (seed Self complementary sequences). Default is FALSE.

correction.factor
Correction factor to be used to modulate the effect of the nucleic acid concentration (nucleic.acid.conc) in the computation of melting temperature (see Correction factor for nucleic acid concentration).

method.approx
Specify the approximative formula to be used for melting temperature calculation for sequences of length greater than size.threshold. Either "ahs01", "che93", "che93corr", "schdot", "owe69", "san98", "wetdna91", "wetrna91" or "wetdnarna91" (see Approximative formulas).

method.nn
Specify the nearest neighbor model to be used for melting temperature calculation for perfectly matching sequences of length lesser than size.threshold. Either "all97", "bre86", "san04", "san96", "sug96", "tan04", "fre86", "xia98", "sug95" or "tur06" (see Perfectly matching sequences).

method.GU
Specify the nearest neighbor model to compute the contribution of GU base pairs to the thermodynamic of helix-coil transition. Either "tur99" or "ser12" (see GU wobble base pairs effect).

method.singleMM
Specify the nearest neighbor model to compute the contribution of single mismatch to the thermodynamic of helix-coil transition. Either "allsanpey", "tur06", "zno07" "zno08" or "wat11" (see Single mismatch effect).

method.tandemMM
Specify the nearest neighbor model to compute the contribution of tandem mismatches to the thermodynamic of helix-coil transition. Either "allsanpey" or "tur99" (see Tandem mismatches effect).

method.single.dangle
Specify the nearest neighbor model to compute the contribution of single dangling end to the thermodynamic of helix-coil transition. Either "bom00", "sugdna02", "sugrna02" or "ser08" (see Single dangling end effect).

method.double.dangle
Specify the nearest neighbor model to compute the contribution of double dangling end to the thermodynamic of helix-coil transition. Either "sugdna02", "sugrna02", "ser05" or "ser06" (see Double dangling end effect).

method.long.dangle
Specify the nearest neighbor model to compute the contribution of long dangling end to the thermodynamic of helix-coil transition. Either "sugdna02" or "sugrna02" (see Long dangling end effect).

method.internal.loop
Specify the nearest neighbor model to compute the contribution of internal loop to the thermodynamic of helix-coil transition. Either "san04", "tur06" or "zno07" (see Internal loop effect).

method.single.bulge.loop
Specify the nearest neighbor model to compute the contribution of single bulge loop to the thermodynamic of helix-coil transition. Either "san04", "tan04", "ser07" or "tur06" (see Single bulge loop effect).

method.long.bulge.loop
Specify the nearest neighbor model to compute the contribution of long bulge loop to the thermodynamic of helix-coil transition. Either "san04" or "tur06" (see Long bulge loop effect).

method.CNG
Specify the nearest neighbor model to compute the contribution of CNG repeats to the thermodynamic of helix-coil transition. Available method is "bro05" (see CNG repeats effect).

method.inosine
Specify the specific nearest neighbor model to compute the contribution of inosine bases (I) to the thermodynamic of helix-coil transition. Either "san05" or "zno07" (see Inosine bases effect).

method.hydroxyadenine
Specify the nearest neighbor model to compute the contribution of hydroxyadenine bases (A*) to the thermodynamic of helix-coil transition. Available method is "sug01" (see Hydroxyadenine bases effect).

method.azobenzenes
Specify the nearest neighbor model to compute the contribution of azobenzenes (X_T for trans azobenzenes and X_C for cis azobenzenes) to the thermodynamic of helix-coil transition. Available method is "asa05" (see Azobenzenes effect).

method.locked
Specify the nearest neighbor model to compute the contribution of locked nucleic acids (AL, GL, TL and CL) to the thermodynamic of helix-coil transition. Available method is "mct04" (see Locked nucleic acids effect).

correction.ion
Specify the correction method for ions. Either one of the following:

Mg corrections"owcmg08", "tanmg06" or "tanmg07" (see Magnesium corrections)

Mixed Na Mg corrections"owcmix08", "tanmix07" or "tanmix07" (see Mixed Sodium and Magnesium corrections)

.

method.Naeq
Specify the ion correction which gives a sodium equivalent concentration if other cations are present. Either "ahs01", "mit96" or "pey00" (see Sodium equivalent concentration methods).

correction.DMSO
Specify the correction method for DMSO. Specify the correction method for DMSO. Either "ahs01", "mus81", "cul76" or "esc80" (see DMSO corrections).

correction.formamide
Specify the correction method for formamide. Specify the correction method for formamide Either "bla96" or "lincorr" (see Formamide corrections).

Value

A list with the following components:

Environment

A list with details about the melting temperature computation environment.

Options

A list with details about the options (default or user specified) used for melting temperature computation.

Results

A list with the results of the melting temperature computation including the enthalpy and entropy in case of nearest neighbour methods.

Message

Error and/or Warning messages, if any.

Mandatory arguments

The following are the arguments which are mandatory for computation.

sequence

5' to 3' sequence of one strand of the nucleic acid duplex as a character string. Recognises A, C, G, T, U, I, X_C, X_T, A*, AL, TL, GL and CL. U and T are not considered identical (see Recognized nucleotides).

comp.sequence

Mandatory if there are mismatches, inosine(s) or hydroxyadenine(s) between the two strands. If not specified, it is computed as the complement of sequence. Self-complementarity in sequence is detected even though there may be (are) dangling end(s) and comp.sequence is computed (see Self complementary sequences).

nucleic.acid.conc

See Correction factor for nucleic acid concentration.

Na.conc, Mg.conc, Tris.conc, K.conc

At least one cation (Na, Mg, Tris, K) concentration is mandatory, the other agents(dNTP, DMSO, formamide) are optional (see Ion and agent concentrations).

hybridisation.type

See Hybridisation type options.

Recognized nucleotides

Code

Type

A

Adenine	C

Cytosine	G

Guanine

T	Thymine

U	Uracil

I	Inosine

X_C	Trans

azobenzenes	X_T

Cis azobenzenes	A*

Hydroxyadenine	AL

Locked nucleic acid

TL	"

GL	"

CL	"

U and T are not considered identical.

Hybridisation type options

The details of the possible options for hybridisation type specified in the argument hybridisation.type are as follows:

Option

Sequence

Complementary sequence

dnadna

DNA	DNA

rnarna

RNA	RNA

dnarna

DNA	RNA

rnadna

RNA	DNA

mrnarna		2-o-methyl RNA

RNA

rnamrna		RNA

2-o-methyl RNA		Option

This parameter determines the nature of the sequences in the arguments sequence and comp.sequence.

Ion and agent concentrations

Ion concentrations are specified by the arguments Na.conc, Mg.conc, Tris.conc and K.conc, while agent concentrations are specified by the arguments dNTP.conc, DMSO.conc and formamide.conc.

These values are used for different correction functions which approximately adjusts for effects of these ions (Na, Mg, Tris, K) and/or agents (dNTP, DMSO, formamide) on on thermodynamic stability of nucleic acid duplexes. Their concentration limits depends on the correction method used. All the concentrations must be in M, except for the DMSO (%) and formamide (% or M depending on the correction method). Note that [Tris+] is about half of the total tris buffer concentration.

Self complementary sequences

Self complementarity for perfect matching sequences or sequences with dangling ends is detected automatically. However it can be enforced by the argument force.self = TRUE.

Correction factor for nucleic acid concentration

For self complementary sequences (Auto detected or specified by force.self) it is 1. Otherwise it is 4 if the both strands are present in equivalent amount and 1 if one strand is in excess.

Approximative estimation formulas

Formula

Type

Limits/Remarks

Reference			ahs01

DNA	No mismatch	von Ahsen

et al., 2001		che93

DNA	No mismatch; Na=0, Mg=0.0015,		Marmur and Doty, 1962

Tris=0.01, K=0.05

che93corr

DNA	No mismatch; Na=0, Mg=0.0015,	Marmur and	Doty, 1962

Tris=0.01, K=0.05

schdot

DNA	No mismatch	Wetmur, 1991; Marmur and

Doty,	1962; Chester and

Marshak, 1993; Schildkraut

and Lifson, 1965; Wahl et

al., 1987; Britten et	al.,

1974; Hall et al., 1980		owe69

DNA		No mismatch

Owen et al., 1969;

Frank-Kamenetskii,	1971;

Blake, 1996; Blake and

Delcourt, 1998

san98		DNA	No mismatch

SantaLucia,	1998; von Ahsen

et al., 2001		wetdna91

*		DNA

Wetmur, 1991		wetrna91	*

RNA		Wetmur,	1991

wetdnarna91	*	DNA/RNA

Wetmur, 1991

* Default formula for computation.

Note that calculation is increasingly incorrect when the length of the duplex decreases. Further, it does not take into account nucleic acid concentration.

Nearest neighbor models

Perfectly matching sequences

Model

Type

Limits/Remarks			Reference

all97

*	DNA		Allawi and SantaLucia, 1997

tur06	*

2'-O-MeRNA/	A sodium correction	Kierzek et al.,	2006

RNA	(	san04

) is

automatically	applied to

convert the entropy (Na =

0.1M) into the entropy (Na =

1M).

bre86

DNA		Breslauer et al., 1986

san04		DNA

SantaLucia and Hicks, 2004		san96

DNA		SantaLucia et

al., 1996		sug96

DNA		Sugimoto et al., 1996

tan04		DNA

Tanaka et al., 2004		fre86

RNA

Freier et al., 1986		xia98	*

RNA		Xia et	al., 1998

sug95	*	DNA/

SantaLucia et al., 1996

RNA

* Default model for computation.

GU wobble base pairs effect

Model

Type

Limits/Remarks			Reference

tur99

RNA		Mathews et al., 1999

ser12	*

RNA		Chen et al., 2012

* Default model for computation. GU base pairs are not taken into account by the approximative mode.

Single mismatch effect

Model

Type

Limits.Remarks			Reference

allsanpey

*	DNA		Allawi and SantaLucia, 1997;

Allawi and SantaLucia, 1998;

Allawi and	SantaLucia, 1998;

Allawi and SantaLucia, 1998;

Peyret et al., 1999

wat11	*	DNA/RNA

Watkins et al., 2011		tur06

RNA		Lu et al., 2006

zno07	*

RNA		Davis and Znosko, 2007

zno08		RNA

At least one adjacent GU base	Davis and	Znosko, 2008

pair.

* Default model for computation. Single mismatches are not taken into account by the approximative mode.

Tandem mismatches effect

Model

Type

Limits.Remarks			Reference

allsanpey

*	DNA	Only GT mismatches and TA/TG	Allawi

and SantaLucia, 1997;			mismatches.

Allawi and SantaLucia,	1998;

Allawi and SantaLucia, 1998;

Allawi and SantaLucia, 1998;

Peyret et al., 1999

tur99	*	RNA

No adjacent GU or UG base	Mathews et al.,	1999; Lu et

pairs.	al., 2006

* Default model for computation. Tandem mismatches are not taken into account by the approximative mode. Note that not all the mismatched Crick's pairs have been investigated.

Single dangling end effect

Model

Type

Limits.Remarks			Reference

bom00

*	DNA		Bommarito et al., 2000

sugdna02

DNA	Only terminal poly A self	Ohmichi et al., 2002

complementary sequences.

sugrna02		RNA

Only	terminal poly A self	Ohmichi et al., 2002

complementary	sequences.

ser08	*	RNA

Only 3' UA, GU and UG		O'Toole et al., 2006; Miller

terminal base pairs only 5'		et al., 2008

UG and GU terminal base

pairs.

Model

* Default model for computation. Single dangling ends are not taken into account by the approximative mode.

Double dangling end effect

Model

Type

Limits/Remarks			Reference

sugdna02

*	DNA	Only terminal poly A self	Ohmichi et al.,

2002			complementary sequences.

sugrna02

RNA	Only terminal poly A self	Ohmichi et al., 2002

complementary sequences.

ser05

RNA	Depends on the	available	O'Toole et al., 2005

thermodynamic parameters	for

single dangling end.

ser06	*

RNA		O'Toole et al., 2006

* Default model for computation. Double dangling ends are not taken into account by the approximative mode.

Long dangling end effect

Model

Type

Limits/Remarks			Reference

sugdna02

*	DNA	Only terminal poly A self	Ohmichi et

al., 2002			complementary sequences.

sugrna02	*

RNA	Only terminal poly A self	Ohmichi et al., 2002

complementary sequences.

Model

* Default model for computation. Long dangling ends are not taken into account by the approximative mode.

Internal loop effect

Model

Type

Limits.Remarks			Reference

san04

*	DNA	Missing asymmetry penalty.	SantaLucia and

Hicks, 2004			Not tested with experimental

results.			tur06

RNA	Not tested with experimental

Lu et al., 2006			results.

zno07	*

RNA		Only for 1x2 loop.	Badhwar et al., 2007

* Default model for computation. Internal loops are not taken into account by the approximative mode.

Single bulge loop effect

Model

Type

Limits/Remarks			Reference

tan04

*	DNA		Tan and Chen, 2007

san04

DNA	Missing closing AT penalty.	SantaLucia and Hicks, 2004

ser07		RNA

Less reliable results. Some	Blose et al.,	2007

missing parameters.

tur06	*	RNA

Lu et al., 2006		Model

* Default model for computation. Single bulge loops are not taken into account by the approximative mode.

Long bulge loop effect

Model

Type

Limits.Remarks			Reference

san04

*	DNA	Missing closing AT penalty.	SantaLucia and

Hicks, 2004		tur06	*

RNA	Not tested with experimental		Mathews et al., 1999; Lu et

results.	al., 2006

* Default model for computation. Long bulge loops are not taken into account by the approximative mode.

CNG repeats effect

Model

Type

Limits/Remarks			Reference

bro05

*	RNA	Self complementary sequences.	Broda et al.,

2005			2 to 7 CNG repeats.

Model

* Default model for computation. CNG repeats are not taken into account by the approximative mode. The contribution of CNG repeats to the thermodynamic of helix-coil transition can be computed only for 2 to 7 CNG repeats. N represents a single mismatch of type N/N.

Inosine bases effect

Model

Type

Limits/Remarks			Reference

san05

*	DNA	Missing parameters for tandem	Watkins and

SantaLucia, 2005			base pairs containing inosine

bases.			zno07

*	RNA	Only IU base pairs.

Wright et al., 2007		Model

* Default model for computation. Inosine bases (I) are not taken into account by the approximative mode.

Hydroxyadenine bases effect

Model

Type

Limits/Remarks			Reference

sug01

*	DNA	Only 5' GA*C 3'and 5' TA*A 3'	Kawakami et

al., 2001			contexts.

Model

* Default model for computation. Hydroxyadenine bases (A*) are not taken into account by the approximative mode.

Azobenzenes effect effect

Model

Type

Limits/Remarks			Reference

asa05

*	DNA	Less reliable results when	Asanuma et al.,

2005			the number of cis azobenzene

increases.			Model

* Default model for computation. Azobenzenes (X_T for trans azobenzenes and X_C for cis azobenzenes) are not taken into account by the approximative mode.

Locked nucleic acids effect

Model

Type

Limits/Remarks			Reference

mct04

*	DNA		McTigue et al., 2004

* Default model for computation. Locked nucleic acids (AL, GL, TL and CL) are not taken into account by the approximative mode.
Ion corrections

Sodium corrections

Correction

Type

Limits.Remarks			Reference

ahs01

DNA	Na>0.	von Ahsen et al., 2001

schlif

DNA	Na>=0.07; Na<=0.12.	Schildkraut and	Lifson, 1965

tanna06		DNA

Na>=0.001; Na<=1.	Tan and	Chen, 2006

tanna07	*	RNA	Na>=0.003; Na<=1.

Tan and	Chen, 2007		or

2'-O-MeRNA/RNA

wet91		RNA,	Na>0.

Wetmur, 1991		DNA

and

RNA/DNA

kam71

DNA		Na>0; Na>=0.069; Na<=1.02.

Frank-Kamenetskii, 1971			marschdot

DNA	Na>=0.069; Na<=1.02.	Marmur and Doty,

1962; Blake

and Delcourt, 1998		owc1904

DNA	Na>0. (equation 19)	Owczarzy et al., 2004

owc2004

DNA	Na>0. (equation 20)	Owczarzy et al., 2004

owc2104		DNA

Na>0. (equation 21)	Owczarzy et al., 2004

owc2204	*	DNA	Na>0. (equation 22)

Owczarzy et al.,	2004		san96

DNA	Na>=0.1.	SantaLucia et al., 1996

san04

DNA	Na>=0.05; Na<=1.1;	SantaLucia and	Hicks, 2004;

Oligonucleotides inferior to	SantaLucia,

1998			16 bases.

Correction

* Default correction method for computation.

Magnesium corrections

Correction

Type

Limits/Remarks			Reference

owcmg08

*	DNA	Mg>=0.0005; Mg<=0.6.	Owczarzy et al.,

2008		tanmg06

DNA	Mg>=0.0001; Mg<=1; Oligomer	Tan	and Chen, 2006

length superior to 6 base

pairs.

tanmg07	*

RNA	Mg>=0.1; Mg<=0.3.	Tan and	Chen, 2007

* Default correction method for computation.

Mixed Sodium and Magnesium corrections

Correction

Type

Limits.Remarks

Reference			owcmix08

*	DNA	Mg>=0.0005; Mg<=0.6;

Owczarzy et al., 2008			Na+K+Tris/2>0.

tanmix07

DNA,	Mg>=0.1; Mg<=0.3;	Tan and Chen, 2007

RNA	Na+K+Tris/2>=0.1;

or	Na+K+Tris/2<=0.3.

2'-O-MeRNA/RNA

* Default correction method for computation.

The ion correction by Owczarzy et al. (2008) is used by default according to the [Mg2+]0.5 ⁄ [Mon+] ratio, where [Mon+] = [Na+] + [Tris+] + [K+] .

If,

[Mon+] = 0

Default sodium correction is used.

Ratio < 0.22,

Default sodium correction is used.

0.22 <= Ratio < 6

Default mixed Na and Mg correction is used.

Ratio >= 6

Default magnesium correction is used.

Note that [Tris+] is about half of the total tris buffer concentration.

Sodium equivalent concentration methods

Correction

Type

Limits/Remarks

Reference			ahs01

*	DNA		von Ahsen et al.,

2001		mit96

DNA		Mitsuhashi, 1996

pey00			DNA

Peyret, 2000

* Default correction method for computation. For the other types of hybridization, the DNA default correction is used. If there are other cations when an approximative approach is used, a sodium equivalence is automatically computed. In case of nearest neighbor approach, the sodium equivalence will be used only if a sodium correction is specified by the argument correction.ion.
Denaturing agent corrections

DMSO corrections

Correction

Type

Limits/Remarks			Reference

ahs01*

DNA	Not tested with experimental	von Ahsen et

al., 2001			results.

cul76

DNA	Not	tested with experimental	Cullen and Bick, 1976

results.

esc80

DNA	Not tested with experimental	Escara	and Hutton, 1980

results.

mus81		DNA

Not tested with experimental	Musielski et al., 1981

results.

* Default correction method for computation. For the other types of hybridization, the DNA default correction is used. If there is DMSO when an approximative approach is used, a DMSO correction is automatically computed. In case of nearest neighbor approach and approximative approach, the DMSO correction will be used only if a sodium correction is specified by the argument correction.ion.

Formamide corrections

Correction

Type

Limits/Remarks			Reference

bla96*

DNA	With formamide concentration	Blake, 1996

in mol/L.

lincorr

DNA	With a

formamide volume.	McConaughy et al., 1969;

Record,	1967; Casey and

Davidson, 1977; Hutton, 1977

* Default correction method for computation. For the other types of hybridization, the DNA default correction is used. If there is formamide when an approximative approach is used, a formamide correction is automatically computed. In case of nearest neighbor approach and approximative approach, the formamide correction will be used only if a sodium correction is specified by the argument correction.ion.
References

Marmur J, Doty P (1962). “Determination of the base composition of deoxyribonucleic acid from its thermal denaturation temperature.” Journal of Molecular Biology, 5(1), 109--118.

Schildkraut C, Lifson S (1965). “Dependence of the melting temperature of DNA on salt concentration.” Biopolymers, 3(2), 195--208.

Record MT (1967). “Electrostatic effects on polynucleotide transitions. I. Behavior at neutral pH.” Biopolymers, 5(10), 975--992.

McConaughy BL, Laird C, McCarthy BJ (1969). “Nucleic acid reassociation in formamide.” Biochemistry, 8(8), 3289--3295.

Owen R, Hill L, Lapage S (1969). “Determination of DNA base compositions from melting profiles in dilute buffers.” Biopolymers, 7(4), 503--516.

Frank-Kamenetskii MD (1971). “Simplification of the empirical relationship between melting temperature of DNA, its GC content and concentration of sodium ions in solution.” Biopolymers, 10(12), 2623--2624.

Britten RJ, Graham DE, Neufeld BR (1974). “Analysis of repeating DNA sequences by reassociation.” Methods in Enzymology, 29, 363--418.

Cullen BR, Bick MD (1976). “Thermal denaturation of DNA from bromodeoxyuridine substituted cells.” Nucleic Acids Research, 3(1), 49--62.

Hutton JR (1977). “Renaturation kinetics and thermal stability of DNA in aqueous solutions of formamide and urea.” Nucleic Acids Research, 4(10), 3537--3555.

Casey J, Davidson N (1977). “Rates of formation and thermal stabilities of RNA:DNA and DNA:DNA duplexes at high concentrations of formamide.” Nucleic Acids Research, 4(5), 1539--1552.

Hall TJ, Grula JW, Davidson EH, Britten RJ (1980). “Evolution of sea urchin non-repetitive DNA.” Journal of Molecular Evolution, 16(2), 95--110.

Escara JF, Hutton JR (1980). “Thermal stability and renaturation of DNA in dimethyl sulfoxide solutions: Acceleration of the renaturation rate.” Biopolymers, 19(7), 1315--1327.

Musielski H, Mann W, Laue R, Michel S (1981). “Influence of dimethylsulfoxide on transcription by bacteriophage T3-induced RNA polymerase.” Zeitschrift fur allgemeine Mikrobiologie, 21(6), 447--456.

Freier SM, Kierzek R, Jaeger JA, Sugimoto N, Caruthers MH, Neilson T, Turner DH (1986). “Improved free-energy parameters for predictions of RNA duplex stability.” Proceedings of the National Academy of Sciences, 83(24), 9373.

Breslauer KJ, Frank R, Blocker H, Marky LA (1986). “Predicting DNA duplex stability from the base sequence.” Proceedings of the National Academy of Sciences, 83(11), 3746.

Wahl GM, Barger SL, Kimmel AR (1987). “Molecular hybridization of immobilized nucleic acids: Theoretical concepts and practical considerations.” Methods in Enzymology, 152, 399--407.

Wetmur JG (1991). “DNA probes: Applications of the principles of nucleic acid hybridization.” Critical Reviews in Biochemistry and Molecular Biology, 26(3-4), 227--259.

Chester N, Marshak D (1993). “Dimethyl sulfoxide-mediated primer Tm reduction: A method for analyzing the role of renaturation temperature in the polymerase chain reaction.” Analytical Biochemistry, 209(2), 284--290.

Sugimoto N, Katoh M, Nakano S, Ohmichi T, Sasaki M (1994). “RNA/DNA hybrid duplexes with identical nearest-neighbor base-pairs have identical stability.” FEBS Letters, 354(1), 74--78.

Sugimoto N, Nakano S, Katoh M, Matsumura A, Nakamuta H, Ohmichi T, Yoneyama M, Sasaki M (1995). “Thermodynamic parameters to predict stability of RNA/DNA hybrid duplexes.” Biochemistry, 34(35), 11211--11216.

SantaLucia J, Allawi HT, Seneviratne PA (1996). “Improved nearest-neighbor parameters for predicting DNA duplex stability.” Biochemistry, 35(11), 3555--3562.

Sugimoto N, Nakano S, Yoneyama M, Honda K (1996). “Improved thermodynamic parameters and helix initiation factor to predict stability of DNA duplexes.” Nucleic Acids Research, 24(22), 4501--4505.

Blake RD, Delcourt SG (1996). “Thermodynamic effects of formamide on DNA stability.” Nucleic Acids Research, 24(11), 2095--2103.

Blake RD (1996). “Denaturation of DNA.” In Meyers RA (ed.), Encyclopedia of molecular biology and molecular medicine, volume 2, 1--19. VCH Verlagsgesellschaft, Weinheim, Germany.

Mitsuhashi M (1996). “Technical report: Part 1. Basic requirements for designing optimal oligonucleotide probe sequences.” Journal of Clinical Laboratory Analysis, 10(5), 277--284.

Allawi HT, SantaLucia J (1997). “Thermodynamics and NMR of internal G·T mismatches in dna.” Biochemistry, 36(34), 10581--10594.

SantaLucia J (1998). “A unified view of polymer, dumbbell, and oligonucleotide DNA nearest-neighbor thermodynamics.” Proceedings of the National Academy of Sciences, 95(4), 1460.

Xia T, SantaLucia J, Burkard ME, Kierzek R, Schroeder SJ, Jiao X, Cox C, Turner DH (1998). “Thermodynamic parameters for an expanded nearest-neighbor model for formation of RNA duplexes with Watson-Crick base pairs.” Biochemistry, 37(42), 14719--14735.

Allawi HT, SantaLucia J (1998). “Thermodynamics of internal C·T mismatches in DNA.” Nucleic Acids Research, 26(11), 2694--2701.

Blake RD, Delcourt SG (1998). “Thermal stability of DNA.” Nucleic Acids Research, 26(14), 3323--3332.

Allawi HT, SantaLucia J (1998). “Nearest neighbor thermodynamic parameters for internal G·A mismatches in DNA.” Biochemistry, 37(8), 2170--2179.

Allawi HT, SantaLucia J (1998). “Nearest-neighbor thermodynamics of internal A·C mismatches in dna: sequence dependence and pH effects.” Biochemistry, 37(26), 9435--9444.

Mathews DH, Sabina J, Zuker M, Turner DH (1999). “Expanded sequence dependence of thermodynamic parameters improves prediction of RNA secondary structure.” Journal of Molecular Biology, 288(5), 911--940.

Peyret N, Seneviratne PA, Allawi HT, SantaLucia J (1999). “Nearest-Neighbor Thermodynamics and NMR of DNA Sequences with Internal A·A, C·C, G·G, and T·T Mismatches.” Biochemistry, 38(12), 3468--3477.

Peyret N (2000). Prediction of nucleic acid hybridization: Parameters and algorithms. Ph.D. Thesis, Wayne State University, Detroit, MI.

Bommarito S, Peyret N, SantaLucia J (2000). “Thermodynamic parameters for DNA sequences with dangling ends.” Nucleic Acids Research, 28(9), 1929--1934.

Kawakami J, Kamiya H, Yasuda K, Fujiki H, Kasai H, Sugimoto N (2001). “Thermodynamic stability of base pairs between 2-hydroxyadenine and incoming nucleotides as a determinant of nucleotide incorporation specificity during replication.” Nucleic Acids Research, 29(16), 3289--3296.

von Ahsen N, Wittwer CT, Schutz E (2001). “Oligonucleotide melting temperatures under PCR conditions: Nearest-neighbor corrections for Mg2+, deoxynucleotide triphosphate, and dimethyl sulfoxide concentrations with comparison to alternative empirical formulas.” Clinical Chemistry, 47(11), 1956--1961.

Le Novere N (2001). “MELTING, computing the melting temperature of nucleic acid duplex.” Bioinformatics, 17(12), 1226--1227.

Ohmichi T, Nakano S, Miyoshi D, Sugimoto N (2002). “Long RNA dangling end has large energetic contribution to duplex stability.” Journal of the American Chemical Society, 124(35), 10367--10372.

SantaLucia J, Hicks D (2004). “The thermodynamics of DNA structural motifs.” Annual Review of Biophysics and Biomolecular Structure, 33(1), 415--440.

Tanaka F, Kameda A, Yamamoto M, Ohuchi A (2004). “Thermodynamic parameters based on a nearest-neighbor model for DNA sequences with a single-bulge loop.” Biochemistry, 43(22), 7143--7150.

McTigue PM, Peterson RJ, Kahn JD (2004). “Sequence-dependent thermodynamic parameters for locked nucleic acid (LNA)-DNA duplex formation.” Biochemistry, 43(18), 5388--5405.

Owczarzy R, You Y, Moreira BG, Manthey JA, Huang L, Behlke MA, Walder JA (2004). “Effects of sodium ions on DNA duplex oligomers: Improved predictions of melting temperatures.” Biochemistry, 43(12), 3537--3554.

Broda M, Kierzek E, Gdaniec Z, Kulinski T, Kierzek R (2005). “Thermodynamic stability of RNA structures formed by CNG trinucleotide repeats. Implication for prediction of RNA structure.” Biochemistry, 44(32), 10873--10882.

Watkins NE, SantaLucia J (2005). “Nearest-neighbor thermodynamics of deoxyinosine pairs in DNA duplexes.” Nucleic Acids Research, 33(19), 6258--6267.

Asanuma H, Matsunaga D, Komiyama M (2005). “Clear-cut photo-regulation of the formation and dissociation of the DNA duplex by modified oligonucleotide involving multiple azobenzenes.” Nucleic Acids Symposium Series, 35--36.

O'Toole AS, Miller S, Serra MJ (2005). “Stability of 3' double nucleotide overhangs that model the 3' ends of siRNA.” RNA, 11(4), 512--516.

Lu ZJ, Turner DH, Mathews DH (2006). “A set of nearest neighbor parameters for predicting the enthalpy change of RNA secondary structure formation.” Nucleic Acids Research, 34(17), 4912--4924.

Kierzek E, Mathews DH, Ciesielska A, Turner DH, Kierzek R (2006). “Nearest neighbor parameters for Watson-Crick complementary heteroduplexes formed between 2'-O-methyl RNA and RNA oligonucleotides.” Nucleic Acids Research, 34(13), 3609--3614.

Tan Z, Chen S (2006). “Nucleic acid helix stability: Effects of salt concentration, cation valence and size, and chain length.” Biophysical Journal, 90(4), 1175--1190.

O'Toole AS, Miller S, Haines N, Zink MC, Serra MJ (2006). “Comprehensive thermodynamic analysis of 3' double-nucleotide overhangs neighboring Watson-Crick terminal base pairs.” Nucleic Acids Research, 34(11), 3338--3344.

Tan Z, Chen S (2007). “RNA helix stability in mixed Na(+)/Mg(2+) solution.” Biophysical Journal, 92(10), 3615--3632.

Wright DJ, Rice JL, Yanker DM, Znosko BM (2007). “Nearest neighbor parameters for inosine·uridine pairs in RNA duplexes.” Biochemistry, 46(15), 4625--4634.

Davis AR, Znosko BM (2007). “Thermodynamic characterization of single mismatches found in naturally occurring RNA.” Biochemistry, 46(46), 13425--13436.

Blose JM, Manni ML, Klapec KA, Stranger-Jones Y, Zyra AC, Sim V, Griffith CA, Long JD, Serra MJ (2007). “Non-nearest-neighbor dependence of stability for RNA bulge loops based on the complete set of group i single nucleotide bulge loops.” Biochemistry, 46(51), 15123--15135.

Badhwar J, Karri S, Cass CK, Wunderlich EL, Znosko BM (2007). “Thermodynamic characterization of RNA duplexes containing naturally occurring 1 * 2 nucleotide internal loops.” Biochemistry, 46(50), 14715--14724.

Davis AR, Znosko BM (2008). “Thermodynamic characterization of naturally occurring RNA single mismatches with G-U nearest neighbors.” Biochemistry, 47(38), 10178--10187.

Miller S, Jones LE, Giovannitti K, Piper D, Serra MJ (2008). “Thermodynamic analysis of 5' and 3' single- and 3' double-nucleotide overhangs neighboring wobble terminal base pairs.” Nucleic Acids Research, 36(17), 5652--5659.

Owczarzy R, Moreira BG, You Y, Behlke MA, Walder JA (2008). “Predicting stability of DNA duplexes in solutions containing magnesium and monovalent cations.” Biochemistry, 47(19), 5336--5353.

Watkins NE, Kennelly WJ, Tsay MJ, Tuin A, Swenson L, Lee H, Morosyuk S, Hicks DA, SantaLucia J (2011). “Thermodynamic contributions of single internal rA·dA, rC·dC, rG·dG and rU·dT mismatches in RNA/DNA duplexes.” Nucleic Acids Research, 39(5), 1894--1902.

Chen JL, Dishler AL, Kennedy SD, Yildirim I, Liu B, Turner DH, Serra MJ (2012). “Testing the nearest neighbor model for canonical rna base pairs: Revision of GU parameters.” Biochemistry, 51(16), 3508--3522.

Dumousseau M, Rodriguez N, Juty N, Le Novere N (2012). “MELTING, a flexible platform to predict the melting temperatures of nucleic acids.” BMC Bioinformatics, 13, 101.

See also

Examples

# Basic usage
melting(sequence = "CAGTGAGACAGCAATGGTCG", nucleic.acid.conc = 2e-06,
        hybridisation.type = "dnadna", Na.conc = 1)

#> [1] 73.35168

# For more detailed examples refer the vignette.

browseVignettes(package = 'rmelting')

#> No vignettes found by browseVignettes(package = "rmelting")